refactor(process): split mod.rs into focused sub-modules
Splits the ~3970-line kernel/src/process/mod.rs into dedicated files:

- task.rs: ProcessControlBlock definition and core methods
- info.rs: ProcessBasicInfo, ThreadInfo, ProcessSignalInfo
- sched_info.rs: ProcessSchedulerInfo, SchedInfo, PiProtected
- state.rs: ProcessState, ExitState, ProcessFlags
- kstack.rs: KernelStack and guard page allocation
- manager/mod.rs: ProcessManager — global process table, init, current, find
- manager/exit.rs: ProcessManager exit/release/group_exit logic
- manager/sched.rs: ProcessManager wakeup/scheduling helpers

mod.rs retains only the module declarations and re-exports,
preserving the external API surface. exec.rs and timer.rs are
adjusted for the new import paths.

This is a pure structural refactoring — no behavioral changes,
no Linux compatibility sematics altered.
