+++ kernel/src/process/mod.rs
@@ -1,87 +1,3 @@
-use core::{
-    fmt,
-    hash::Hash,
-    hint::spin_loop,
-    intrinsics::unlikely,
-    mem::ManuallyDrop,
-    str::FromStr,
-    sync::atomic::{
-        compiler_fence, fence, AtomicBool, AtomicI32, AtomicU32, AtomicU64, AtomicU8, AtomicUsize,
-        Ordering,
-    },
-};
-
-use alloc::{
-    ffi::CString,
-    string::{String, ToString},
-    sync::{Arc, Weak},
-    vec::Vec,
-};
-use cred::INIT_CRED;
-use hashbrown::HashMap;
-use log::{debug, error, info, warn};
-use pid::{alloc_pid, Pid, PidLink, PidType};
-use process_group::Pgid;
-use system_error::SystemError;
-
-use crate::{
-    arch::{
-        cpu::current_cpu_id,
-        ipc::signal::{AtomicSignal, SigFlags, SigSet, Signal},
-        process::ArchPCBInfo,
-        CurrentIrqArch, SigStackArch,
-    },
-    cgroup::{cgroup_root_node, CgroupNode, TaskCgroupRef},
-    driver::tty::{tty_core::TtyCore, tty_job_control::TtyJobCtrlManager},
-    exception::InterruptArch,
-    filesystem::{
-        fs::FsStruct,
-        vfs::{file::FileDescriptorVec, FileType, IndexNode},
-    },
-    ipc::{
-        sighand::SigHand,
-        signal::RestartBlock,
-        signal_types::{SigCode, SigInfo, SigPending, SigType, SignalFlags},
-    },
-    libs::{
-        align::AlignedBox,
-        cpumask::CpuMask,
-        futex::{
-            constant::{FutexFlag, FUTEX_BITSET_MATCH_ANY},
-            futex::{Futex, RobustListHead},
-        },
-        lock_free_flags::LockFreeFlags,
-        mutex::{Mutex, MutexGuard},
-        rwlock::{RwLock, RwLockReadGuard, RwLockUpgradableGuard, RwLockWriteGuard},
-        rwsem::RwSem,
-        spinlock::{SpinLock, SpinLockGuard},
-        wait_queue::WaitQueue,
-    },
-    mm::{
-        percpu::{PerCpu, PerCpuVar},
-        set_IDLE_PROCESS_ADDRESS_SPACE,
-        ucontext::AddressSpace,
-        PhysAddr, VirtAddr, IDLE_PROCESS_ADDRESS_SPACE,
-    },
-    process::resource::{RLimit64, RLimitID, RUsage},
-    rcu::RcuArcSlot,
-    sched::{
-        DequeueFlag, EnqueueFlag, OnRq, SchedMode, SchedPolicy, Scheduler, WakeupFlags,
-        __schedule_with_current, completion::Completion, cpu_is_online, cpu_rq,
-        enqueue_task_on_cpu, fair::FairSchedEntity, prio::MAX_PRIO, select_task_rq,
-    },
-    smp::{
-        core::smp_get_processor_id,
-        cpu::{AtomicProcessorId, ProcessorId},
-        kick_cpu,
-    },
-    syscall::user_access::{clear_user_protected, write_one_to_user_protected},
-};
-use timer::AlarmTimer;
-
-use self::{cred::Cred, kthread::WorkerPrivate};
-use crate::process::namespace::nsproxy::NsProxy;
-
 pub mod abi;
 pub mod cputime;
 pub mod cred;
@@ -111,1280 +27,50 @@ pub mod timer;
 pub mod utils;
 pub mod wait;
 
+use core::sync::atomic::AtomicUsize;
+
+mod info;
+mod kstack;
+mod manager;
+mod sched_info;
+mod state;
+mod task;
+
+int_like!(RawPid, AtomicRawPid, usize, AtomicUsize);
+
 pub use cputime::ProcessCpuTime;
+pub(crate) use cred::Cred;
+#[allow(unused_imports)]
+pub use info::{
+    CpuItimer, ProcessBasicInfo, ProcessItimer, ProcessItimers, ProcessSignalInfo, ThreadInfo,
+};
+#[allow(unused_imports)]
+pub use kstack::{KernelStack, KernelStackType};
+pub(crate) use manager::{
+    account_context_switch, account_successful_fork, all_process, dec_visible_thread_count,
+    inc_visible_thread_count,
+};
+#[allow(unused_imports)]
+pub use manager::{
+    nr_context_switches, nr_threads, total_forks, ProcessManager, SwitchResult,
+    PROCESS_SWITCH_RESULT,
+};
+use manager::{PTRACE_RELATION_LOCK, __PROCESS_MANAGEMENT_INIT_DONE};
+use pid::alloc_pid;
+pub(crate) use process_group::Pgid;
+#[allow(unused_imports)]
+pub(crate) use sched_info::NewTaskPlacement;
+#[allow(unused_imports)]
+pub use sched_info::{PiProtected, ProcessSchedulerInfo, SchedInfo};
+pub use state::{ExitState, ProcessFlags, ProcessState};
+pub use task::ProcessControlBlock;
 
-/// 系统中所有进程的pcb
-static ALL_PROCESS: SpinLock<Option<HashMap<RawPid, Arc<ProcessControlBlock>>>> =
-    SpinLock::new(None);
-static PTRACE_RELATION_LOCK: SpinLock<()> = SpinLock::new(());
-static NR_VISIBLE_THREADS: AtomicUsize = AtomicUsize::new(0);
-static TOTAL_FORKS: AtomicU64 = AtomicU64::new(0);
-static TOTAL_CONTEXT_SWITCHES: AtomicU64 = AtomicU64::new(0);
-
-pub(crate) fn all_process() -> &'static SpinLock<Option<HashMap<RawPid, Arc<ProcessControlBlock>>>>
-{
-    &ALL_PROCESS
+pub fn process_init() {
+    ProcessManager::init();
 }
 
-#[inline]
-pub(crate) fn inc_visible_thread_count() {
-    NR_VISIBLE_THREADS.fetch_add(1, Ordering::Relaxed);
-}
-
-#[inline]
-pub(crate) fn dec_visible_thread_count() {
-    NR_VISIBLE_THREADS.fetch_sub(1, Ordering::Relaxed);
-}
-
-#[inline]
-pub(crate) fn account_successful_fork() {
-    TOTAL_FORKS.fetch_add(1, Ordering::Relaxed);
-}
-
-#[inline]
-pub fn nr_threads() -> u32 {
-    NR_VISIBLE_THREADS.load(Ordering::Relaxed) as u32
-}
-
-#[inline]
-pub fn total_forks() -> u64 {
-    TOTAL_FORKS.load(Ordering::Relaxed)
-}
-
-#[inline]
-pub(crate) fn account_context_switch() {
-    TOTAL_CONTEXT_SWITCHES.fetch_add(1, Ordering::Relaxed);
-}
-
-#[inline]
-pub fn nr_context_switches() -> u64 {
-    TOTAL_CONTEXT_SWITCHES.load(Ordering::Relaxed)
-}
-
-fn exchange_raw_pids_locked(
-    map: &mut HashMap<RawPid, Arc<ProcessControlBlock>>,
-    left: &Arc<ProcessControlBlock>,
-    right: &Arc<ProcessControlBlock>,
-) -> Result<(), SystemError> {
-    let left_pid = left.raw_pid();
-    let right_pid = right.raw_pid();
-    if left_pid == right_pid {
-        return Err(SystemError::EINVAL);
-    }
-    let left_entry = map.remove(&left_pid).ok_or(SystemError::ESRCH)?;
-    let right_entry = map.remove(&right_pid).ok_or(SystemError::ESRCH)?;
-
-    unsafe {
-        left.force_set_raw_pid(right_pid);
-        right.force_set_raw_pid(left_pid);
-    }
-
-    map.insert(right_pid, left_entry);
-    map.insert(left_pid, right_entry);
-    Ok(())
-}
-
-pub static mut PROCESS_SWITCH_RESULT: Option<PerCpuVar<SwitchResult>> = None;
-
-/// 一个只改变1次的全局变量，标志进程管理器是否已经初始化完成
-static mut __PROCESS_MANAGEMENT_INIT_DONE: bool = false;
-
-/// Serializes `/proc/[pid]/oom_score_adj` writes with `CLONE_VM` inheritance.
-///
-/// This mirrors Linux's `oom_adj_mutex` responsibility: a process sharing an
-/// mm with another thread group must not miss a concurrent oom_score_adj update
-/// while it is becoming visible.
-static OOM_SCORE_ADJ_LOCK: Mutex<()> = Mutex::new(());
-
-pub struct SwitchResult {
-    pub prev_pcb: Option<Arc<ProcessControlBlock>>,
-    pub next_pcb: Option<Arc<ProcessControlBlock>>,
-    pub migrate_prev_to: Option<ProcessorId>,
-}
-
-impl SwitchResult {
-    pub fn new() -> Self {
-        Self {
-            prev_pcb: None,
-            next_pcb: None,
-            migrate_prev_to: None,
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct ProcessManager;
-impl ProcessManager {
-    pub fn lock_oom_score_adj() -> MutexGuard<'static, ()> {
-        OOM_SCORE_ADJ_LOCK.lock()
-    }
-
-    fn mm_has_user_tasks(mm: &Arc<AddressSpace>) -> bool {
-        ProcessManager::get_all_processes()
-            .into_iter()
-            .filter_map(ProcessManager::find)
-            .any(|task| {
-                task.basic()
-                    .user_vm()
-                    .is_some_and(|task_mm| task_mm.id() == mm.id() || Arc::ptr_eq(&task_mm, mm))
-            })
-    }
-
-    pub fn is_current(pcb: &Arc<ProcessControlBlock>) -> bool {
-        Arc::ptr_eq(pcb, &Self::current_pcb())
-    }
-
-    pub fn thread_group_leader_of(pcb: &Arc<ProcessControlBlock>) -> Arc<ProcessControlBlock> {
-        pcb.threads_read_irqsave()
-            .group_leader()
-            .unwrap_or_else(|| pcb.clone())
-    }
-
-    /// 遍历 `pcb` 所在线程组中的线程。
-    ///
-    /// 遍历顺序为线程组组长优先，随后遍历组长维护的其他线程列表；
-    /// 回调返回 `false` 时提前停止遍历。
-    pub fn for_each_thread_in_group<F>(pcb: Arc<ProcessControlBlock>, mut func: F)
-    where
-        F: FnMut(Arc<ProcessControlBlock>) -> bool,
-    {
-        let thread_group_leader = Self::thread_group_leader_of(&pcb);
-        if !func(thread_group_leader.clone()) {
-            return;
-        }
-
-        let group_tasks = thread_group_leader
-            .threads_read_irqsave()
-            .group_tasks_clone();
-        for weak in group_tasks {
-            let Some(task) = weak.upgrade() else {
-                continue;
-            };
-            if Arc::ptr_eq(&task, &thread_group_leader) {
-                continue;
-            }
-            if !func(task) {
-                break;
-            }
-        }
-    }
-
-    pub fn thread_group_tasks_snapshot(
-        pcb: Arc<ProcessControlBlock>,
-    ) -> Vec<Arc<ProcessControlBlock>> {
-        let leader = Self::thread_group_leader_of(&pcb);
-        let mut tasks = Vec::new();
-        tasks.push(leader.clone());
-
-        let group_tasks = leader.threads_read_irqsave().group_tasks_clone();
-        for weak in group_tasks {
-            let Some(task) = weak.upgrade() else {
-                continue;
-            };
-            if tasks.iter().any(|existing| Arc::ptr_eq(existing, &task)) {
-                continue;
-            }
-            tasks.push(task);
-        }
-
-        tasks
-    }
-
-    #[inline(never)]
-    fn init() {
-        static INIT_FLAG: AtomicBool = AtomicBool::new(false);
-        if INIT_FLAG
-            .compare_exchange(false, true, Ordering::SeqCst, Ordering::SeqCst)
-            .is_err()
-        {
-            panic!("ProcessManager has been initialized!");
-        }
-
-        unsafe {
-            compiler_fence(Ordering::SeqCst);
-            debug!("To create address space for INIT process.");
-            // test_buddy();
-            set_IDLE_PROCESS_ADDRESS_SPACE(
-                AddressSpace::new(true).expect("Failed to create address space for INIT process."),
-            );
-            debug!("INIT process address space created.");
-            compiler_fence(Ordering::SeqCst);
-        };
-
-        ALL_PROCESS.lock_irqsave().replace(HashMap::new());
-        Self::init_switch_result();
-        Self::arch_init();
-        debug!("process arch init done.");
-        Self::init_idle();
-        debug!("process idle init done.");
-
-        unsafe { __PROCESS_MANAGEMENT_INIT_DONE = true };
-        info!("Process Manager initialized.");
-    }
-
-    fn init_switch_result() {
-        let mut switch_res_vec: Vec<SwitchResult> = Vec::new();
-        for _ in 0..PerCpu::MAX_CPU_NUM {
-            switch_res_vec.push(SwitchResult::new());
-        }
-        unsafe {
-            PROCESS_SWITCH_RESULT = Some(PerCpuVar::new(switch_res_vec).unwrap());
-        }
-    }
-
-    /// 判断进程管理器是否已经初始化完成
-    #[allow(dead_code)]
-    pub fn initialized() -> bool {
-        unsafe { __PROCESS_MANAGEMENT_INIT_DONE }
-    }
-
-    /// 获取当前进程的pcb
-    pub fn current_pcb() -> Arc<ProcessControlBlock> {
-        if unlikely(unsafe { !__PROCESS_MANAGEMENT_INIT_DONE }) {
-            error!("unsafe__PROCESS_MANAGEMENT_INIT_DONE == false");
-            loop {
-                spin_loop();
-            }
-        }
-        return ProcessControlBlock::arch_current_pcb();
-    }
-
-    /// 获取当前进程的pid
-    ///
-    /// 如果进程管理器未初始化完成，那么返回0
-    pub fn current_pid() -> RawPid {
-        if unlikely(unsafe { !__PROCESS_MANAGEMENT_INIT_DONE }) {
-            return RawPid(0);
-        }
-
-        return ProcessManager::current_pcb().raw_pid();
-    }
-
-    /// 根据pid获取进程的pcb
-    ///
-    /// ## 参数
-    ///
-    /// - `pid` : 进程的pid
-    ///
-    /// ## 返回值
-    ///
-    /// 如果找到了对应的进程，那么返回该进程的pcb，否则返回None
-    pub fn find(pid: RawPid) -> Option<Arc<ProcessControlBlock>> {
-        return ALL_PROCESS.lock_irqsave().as_ref()?.get(&pid).cloned();
-    }
-
-    /// 向系统中添加一个进程的pcb
-    ///
-    /// ## 参数
-    ///
-    /// - `pcb` : 进程的pcb
-    ///
-    /// ## 返回值
-    ///
-    /// 无
-    pub fn add_pcb(pcb: Arc<ProcessControlBlock>) {
-        ALL_PROCESS
-            .lock_irqsave()
-            .as_mut()
-            .unwrap()
-            .insert(pcb.raw_pid(), pcb.clone());
-    }
-
-    pub(crate) fn exchange_tid_and_raw_pids(
-        left: &Arc<ProcessControlBlock>,
-        right: &Arc<ProcessControlBlock>,
-    ) -> Result<(), SystemError> {
-        let _cgroup_guard = crate::cgroup::cgroup_accounting_lock().lock();
-        let mut all_proc = all_process().lock_irqsave();
-        let map = all_proc.as_mut().ok_or(SystemError::EINVAL)?;
-        let left_old_pid = left.raw_pid();
-        let right_old_pid = right.raw_pid();
-        if left_old_pid == right_old_pid {
-            return Err(SystemError::EINVAL);
-        }
-        if !map.contains_key(&left_old_pid) || !map.contains_key(&right_old_pid) {
-            return Err(SystemError::ESRCH);
-        }
-
-        let left_cgroup = left.task_cgroup_node();
-        let right_cgroup = right.task_cgroup_node();
-        let left_alive = !left.is_exited();
-        let right_alive = !right.is_exited();
-
-        left.exchange_tid_with(right)?;
-        exchange_raw_pids_locked(map, left, right)?;
-
-        // cgroup.procs 只展示仍存活的任务；exec 去线程化换 pid 时只重命名可见成员。
-        if Arc::ptr_eq(&left_cgroup, &right_cgroup) && left_alive && right_alive {
-            return Ok(());
-        }
-        if left_alive {
-            left_cgroup.rename_task(left_old_pid, left.raw_pid());
-        }
-        if right_alive {
-            right_cgroup.rename_task(right_old_pid, right.raw_pid());
-        }
-
-        Ok(())
-    }
-
-    /// ### 获取所有进程的pid
-    pub fn get_all_processes() -> Vec<RawPid> {
-        let mut pids = Vec::new();
-        for (pid, _) in ALL_PROCESS.lock_irqsave().as_ref().unwrap().iter() {
-            pids.push(*pid);
-        }
-        pids
-    }
-
-    /// 唤醒一个进程
-    pub fn wakeup(pcb: &Arc<ProcessControlBlock>) -> Result<(), SystemError> {
-        let _guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-        let state = pcb.sched_info().state();
-        if !state.is_blocked() {
-            if state.is_exited() {
-                return Err(SystemError::EINVAL);
-            }
-            return Ok(());
-        }
-
-        // 在 pi_lock 保护下读取状态并决定 sched_contributes_to_load
-        let pi_guard = pcb.sched_info().pi_lock_irqsave();
-        fence(Ordering::SeqCst); // smp_mb__after_spinlock()
-        let state = pcb.sched_info().state();
-        if !state.is_blocked() {
-            if state.is_exited() {
-                return Err(SystemError::EINVAL);
-            }
-            return Ok(());
-        }
-        let was_uninterruptible = matches!(state, ProcessState::Blocked(false));
-
-        pcb.sched_info().set_state(ProcessState::Runnable);
-        fence(Ordering::SeqCst);
-
-        pcb.debug_assert_fork_cpu_binding();
-
-        let mut schedule_dequeue_race = false;
-        if *pcb.sched_info().on_rq.lock_irqsave() == OnRq::Queued {
-            if let Some(target_cpu) = pcb.sched_info().on_cpu() {
-                let rq = cpu_rq(target_cpu.data() as usize);
-                let (rq, _rq_guard) = rq.self_lock();
-
-                // Linux ttwu_runnable(): a blocked-but-still-queued task has
-                // not yet been dequeued by schedule(). Recheck on_rq under the
-                // target rq lock; if schedule() won the race and dequeued it,
-                // fall through to the full enqueue path below.
-                if *pcb.sched_info().on_rq.lock_irqsave() == OnRq::Queued {
-                    if !Arc::ptr_eq(&rq.current(), pcb) {
-                        rq.update_rq_clock();
-                        rq.check_preempt_current(pcb, WakeupFlags::WF_TTWU);
-                    }
-                    return Ok(());
-                }
-
-                // DragonOS currently tracks task_cpu, but not Linux's
-                // separate p->on_cpu "still switching out" flag. If schedule()
-                // won the race and just dequeued this task, keep the wakeup on
-                // the previous rq instead of migrating it to another CPU where
-                // the same PCB/kernel stack could run before the old CPU
-                // finishes switch_process().
-                schedule_dequeue_race = true;
-            }
-        }
-
-        let prev_cpu = pcb.sched_info().on_cpu().unwrap_or(current_cpu_id());
-        let allowed = pi_guard.cpus_allowed.clone();
-        let target_cpu = if schedule_dequeue_race {
-            prev_cpu
-        } else {
-            select_task_rq(pcb, prev_cpu, WakeupFlags::WF_TTWU, &allowed)
-        };
-
-        if was_uninterruptible || pcb.flags().contains(ProcessFlags::IN_IOWAIT) {
-            let prev_rq = cpu_rq(prev_cpu.data() as usize);
-            let (prev_rq, _prev_rq_guard) = prev_rq.self_lock();
-            if was_uninterruptible {
-                prev_rq.dec_nr_uninterruptible();
-            }
-            if pcb.flags().contains(ProcessFlags::IN_IOWAIT) {
-                prev_rq.dec_nr_iowait();
-            }
-        }
-
-        enqueue_task_on_cpu(pcb, target_cpu, WakeupFlags::WF_TTWU, false);
-
-        Ok(())
-    }
-
-    // 在 pi_lock 保护下完成状态写入和 CPU 选择
-    pub fn wake_up_new_task(pcb: &Arc<ProcessControlBlock>) -> Result<(), SystemError> {
-        let _guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-
-        debug_assert_eq!(*pcb.sched_info().on_rq.lock_irqsave(), OnRq::None);
-        debug_assert!(pcb.sched_info().is_new_task());
-        debug_assert!(pcb.sched_info().on_cpu().is_none());
-
-        let pi_guard = pcb.sched_info().pi_lock_irqsave();
-        pcb.sched_info().set_state(ProcessState::Runnable);
-
-        let target_cpu = pcb.sched_info().consume_new_task_target_cpu(
-            smp_get_processor_id(),
-            pi_guard.cpus_allowed.clone(),
-            |allowed| {
-                let cpu =
-                    select_task_rq(pcb, smp_get_processor_id(), WakeupFlags::WF_FORK, allowed);
-                if allowed.get(cpu).unwrap_or(false) {
-                    Some(cpu)
-                } else {
-                    None
-                }
-            },
-        )?;
-
-        enqueue_task_on_cpu(pcb, target_cpu, WakeupFlags::WF_FORK, false);
-
-        debug_assert!(!pcb.sched_info().is_new_task());
-        Ok(())
-    }
-
-    /// 将指定内核线程设置为 SCHED_FIFO 调度策略。
-    ///
-    /// task_rq_lock → update_rq_clock → 读取 queued/running →
-    /// dequeue/put_prev → 修改参数 → enqueue → check_class_changed → unlock
-    pub fn set_fifo_policy(pcb: &Arc<ProcessControlBlock>, prio: i32) -> Result<(), SystemError> {
-        if !pcb.flags().contains(ProcessFlags::KTHREAD) {
-            return Err(SystemError::EPERM);
-        }
-
-        if !(0..crate::sched::prio::MAX_RT_PRIO).contains(&prio) {
-            return Err(SystemError::EINVAL);
-        }
-
-        let _irq_guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-
-        // 锁序：pi_lock → rq_lock，与 Linux task_rq_lock() 一致
-        let pi_guard = pcb.sched_info().pi_lock_irqsave();
-
-        let target_cpu = pcb.sched_info().on_cpu().unwrap_or(current_cpu_id());
-        let update_clock = target_cpu == smp_get_processor_id();
-        let rq = cpu_rq(target_cpu.data() as usize);
-        let (rq, rq_guard) = rq.self_lock();
-        if update_clock {
-            rq.update_rq_clock();
-        }
-
-        // 在锁内读取任务状态
-        let old_policy = pcb.sched_info().policy();
-        let queued = *pcb.sched_info().on_rq.lock_irqsave() == OnRq::Queued;
-
-        // 判断目标是否就是 rq 上正在运行的任务
-        let running = Arc::ptr_eq(&rq.current(), pcb);
-
-        // 先将任务从调度器中摘出，再修改参数，最后重新入队
-        if queued {
-            rq.dequeue_task(
-                pcb.clone(),
-                DequeueFlag::DEQUEUE_NOCLOCK | DequeueFlag::DEQUEUE_SAVE,
-            );
-        }
-
-        // running 任务需要先 put_prev_task 让出当前执行位置
-        if running {
-            match old_policy {
-                SchedPolicy::FIFO => {
-                    crate::sched::fifo::FifoScheduler::put_prev_task(rq, pcb.clone())
-                }
-                SchedPolicy::CFS => {
-                    crate::sched::fair::CompletelyFairScheduler::put_prev_task(rq, pcb.clone())
-                }
-                SchedPolicy::IDLE => {
-                    crate::sched::idle::IdleScheduler::put_prev_task(rq, pcb.clone())
-                }
-                SchedPolicy::RT => todo!("RT scheduler not implemented yet"),
-            }
-        }
-
-        // 修改调度参数（rq_lock 保护下）
-        // 对标 Linux __setscheduler_params + __setscheduler_prio：
-        //   - policy 设为 FIFO
-        //   - prio = normal_prio = MAX_RT_PRIO - 1 - rt_priority（调用者传入的已是内核 prio）
-        //   - static_prio 不动（Linux 只对 fair_policy 修改 static_prio，core.c:7528-7529）
-        pcb.sched_info().set_policy(SchedPolicy::FIFO);
-        pcb.sched_info().set_prio(prio);
-        pcb.sched_info().set_normal_prio(prio);
-
-        // 重新入队
-        if queued {
-            rq.enqueue_task(
-                pcb.clone(),
-                EnqueueFlag::ENQUEUE_NOCLOCK | EnqueueFlag::ENQUEUE_RESTORE,
-            );
-        }
-
-        // 对标 Linux __sched_setscheduler: running 任务修改策略后需要 set_next_task
-        if running {
-            match pcb.sched_info().policy() {
-                SchedPolicy::FIFO => {
-                    crate::sched::fifo::FifoScheduler::set_next_task(rq, pcb.clone());
-                }
-                SchedPolicy::CFS => {
-                    crate::sched::fair::CompletelyFairScheduler::set_next_task(rq, pcb.clone());
-                }
-                _ => {}
-            }
-        }
-
-        // check_class_changed → 抢占检查
-        if update_clock {
-            rq.check_preempt_current(pcb, WakeupFlags::empty());
-        } else {
-            rq.check_preempt_remote(pcb, WakeupFlags::empty());
-        }
-
-        // 释放顺序：先 rq_lock 再 pi_lock，与 Linux task_rq_unlock() 一致
-        drop(rq_guard);
-        drop(pi_guard);
-
-        Ok(())
-    }
-
-    /// 唤醒暂停的进程
-    pub fn wakeup_stop(pcb: &Arc<ProcessControlBlock>) -> Result<(), SystemError> {
-        let _guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-        let state = pcb.sched_info().state();
-        if !state.is_stopped() {
-            return if state.is_runnable() {
-                Ok(())
-            } else {
-                Err(SystemError::EINVAL)
-            };
-        }
-
-        let _pi_guard = pcb.sched_info().pi_lock_irqsave();
-        let state = pcb.sched_info().state();
-        if !state.is_stopped() {
-            return if state.is_runnable() {
-                Ok(())
-            } else {
-                Err(SystemError::EINVAL)
-            };
-        }
-
-        pcb.sched_info().set_state(ProcessState::Runnable);
-        fence(Ordering::SeqCst);
-
-        // TODO: select_task_rq 负载均衡
-        // let prev_cpu = pcb.sched_info().on_cpu().unwrap_or(smp_get_processor_id());
-        // let allowed = pi_guard.cpus_allowed.clone();
-        // let target_cpu = select_task_rq(pcb, prev_cpu, WakeupFlags::WF_TTWU, &allowed);
-        let target_cpu = pcb.sched_info().on_cpu().unwrap_or(current_cpu_id());
-        let update_clock = target_cpu == smp_get_processor_id();
-        let rq = cpu_rq(target_cpu.data() as usize);
-
-        let should_enqueue = {
-            let (rq, _rq_guard) = rq.self_lock();
-            match *pcb.sched_info().on_rq.lock_irqsave() {
-                OnRq::Queued => {
-                    let is_current = Arc::ptr_eq(&rq.current(), pcb);
-                    if !is_current {
-                        if update_clock {
-                            rq.update_rq_clock();
-                            rq.check_preempt_current(pcb, WakeupFlags::empty());
-                        } else {
-                            rq.check_preempt_remote(pcb, WakeupFlags::empty());
-                        }
-                    } else if !update_clock {
-                        kick_cpu(target_cpu).ok();
-                    }
-                    false
-                }
-                OnRq::None => true,
-                OnRq::Migrating => false,
-            }
-        };
-
-        if should_enqueue {
-            enqueue_task_on_cpu(pcb, target_cpu, WakeupFlags::empty(), false);
-        }
-
-        Ok(())
-    }
-
-    /// 异步将目标进程置为停止状态（用于 SIGSTOP/SIGTSTP 等作业控制停止）
-    ///
-    /// 注意：该函数用于对“目标进程”进行停止标记，不要求在目标进程上下文调用。
-    /// 与 `mark_stop`（仅当前进程）相对应。
-    ///
-    /// Linux 中 stop 由目标线程在自身上下文中完成（设 Stopped + schedule → deactivate）。
-    /// DragonOS 采用异步方式，因此需要在这里主动 dequeue 排队中的任务，
-    /// 防止 `pick_next_task()` 选中 state=Stopped 的任务短暂运行。
-    ///
-    /// 锁序：pi_lock → rq_lock，与 wakeup() / set_fifo_policy() 序列化一致。
-    pub fn stop_task(pcb: &Arc<ProcessControlBlock>) -> Result<(), SystemError> {
-        let _guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-
-        // 对齐 Linux set_special_state(TASK_STOPPED)：pi_lock 保护状态写入与 exit 检查，
-        // 序列化与 wakeup()/wakeup_stop()/do_exit() 中 pi_lock 保护的并发操作。
-        //
-        // Linux 中 stop 是同步的：目标线程在 get_signal() → do_signal_stop() 中
-        // 调 set_special_state(TASK_STOPPED) + schedule()。
-        // DragonOS 采用异步方式（从发送方上下文停止目标线程），因此需要主动 dequeue
-        // 并 kick 远端 CPU。这是与 Linux 的架构差异，但锁序和 dequeue 语义对标。
-        let pi_guard = pcb.sched_info().pi_lock_irqsave();
-        let prev_state = pcb.sched_info().state();
-        if prev_state.is_exited() {
-            return Err(SystemError::EINTR);
-        }
-        let target_cpu = pcb.sched_info().on_cpu().unwrap_or_else(current_cpu_id);
-        let update_clock = target_cpu == smp_get_processor_id();
-        let was_off_rq = *pcb.sched_info().on_rq.lock_irqsave() == OnRq::None;
-        let was_uninterruptible = matches!(prev_state, ProcessState::Blocked(false));
-        let was_iowait = pcb.flags().contains(ProcessFlags::IN_IOWAIT);
-
-        if was_off_rq && (was_uninterruptible || was_iowait) {
-            let prev_rq = cpu_rq(target_cpu.data() as usize);
-            let (prev_rq, _prev_rq_guard) = prev_rq.self_lock();
-            if was_uninterruptible {
-                prev_rq.dec_nr_uninterruptible();
-            }
-            if was_iowait {
-                prev_rq.dec_nr_iowait();
-            }
-        }
-
-        pcb.sched_info().set_state(ProcessState::Stopped);
-        pcb.flags().insert(ProcessFlags::NEED_SCHEDULE);
-
-        let rq = cpu_rq(target_cpu.data() as usize);
-
-        // 锁序 pi_lock → rq_lock，与 wakeup() 一致
-        let (rq, rq_guard) = rq.self_lock();
-        if update_clock {
-            rq.update_rq_clock();
-        }
-
-        let is_current = Arc::ptr_eq(&rq.current(), pcb);
-
-        if *pcb.sched_info().on_rq.lock_irqsave() == OnRq::Queued {
-            // 排队中且非current → 主动 dequeue
-            if !is_current {
-                rq.deactivate_task(
-                    pcb.clone(),
-                    DequeueFlag::DEQUEUE_STOPPED | DequeueFlag::DEQUEUE_NOCLOCK,
-                );
-            } else if !update_clock {
-                // 远端 CPU 上的 current：只设状态 + kick，不在发送端 dequeue。
-                //
-                // 对标 Linux signal_wake_up_state() + kick_process()：
-                //   发送端仅设 TIF_SIGPENDING / NEED_SCHEDULE 并 kick，不做 dequeue。
-                //   远端 CPU 的 __schedule() 看到非 runnable 的 prev 后自行完成唯一一次出队。
-                //
-                // 现在由 __schedule_inner 统一出队路径 + on_rq 守卫保证幂等。
-                kick_cpu(target_cpu).ok();
-            }
-        }
-
-        drop(rq_guard);
-        drop(pi_guard);
-        Ok(())
-    }
-
-    /// 标志当前进程永久睡眠，但是发起调度的工作，应该由调用者完成
-    ///
-    /// ## 注意
-    ///
-    /// - 进入当前函数之前，不能持有sched_info的锁
-    /// - 进入当前函数之前，必须关闭中断
-    /// - 进入当前函数之后必须保证逻辑的正确性，避免被重复加入调度队列
-    pub fn mark_sleep(interruptable: bool) -> Result<(), SystemError> {
-        assert!(
-            !CurrentIrqArch::is_irq_enabled(),
-            "interrupt must be disabled before enter ProcessManager::mark_sleep()"
-        );
-        let pcb = ProcessManager::current_pcb();
-        if !pcb.sched_info().state().is_exited() {
-            pcb.sched_info()
-                .set_state(ProcessState::Blocked(interruptable));
-            pcb.flags().insert(ProcessFlags::NEED_SCHEDULE);
-            fence(Ordering::SeqCst);
-            return Ok(());
-        }
-        return Err(SystemError::EINTR);
-    }
-
-    /// 标志当前进程为停止状态，但是发起调度的工作，应该由调用者完成
-    ///
-    /// ## 注意
-    ///
-    /// - 进入当前函数之前，不能持有sched_info的锁
-    /// - 进入当前函数之前，必须关闭中断
-    pub fn mark_stop() -> Result<(), SystemError> {
-        assert!(
-            !CurrentIrqArch::is_irq_enabled(),
-            "interrupt must be disabled before enter ProcessManager::mark_stop()"
-        );
-
-        let pcb = ProcessManager::current_pcb();
-        if !pcb.sched_info().state().is_exited() {
-            // pi_lock 保护 STOPPED 写入，序列化与 wakeup()/wakeup_stop() 的并发。
-            let _pi_guard = pcb.sched_info().pi_lock_irqsave();
-            pcb.sched_info().set_state(ProcessState::Stopped);
-            pcb.flags().insert(ProcessFlags::NEED_SCHEDULE);
-            return Ok(());
-        }
-        return Err(SystemError::EINTR);
-    }
-
-    /// 当子进程退出后向父进程发送通知
-    #[inline(never)]
-    fn exit_notify(current: &Arc<ProcessControlBlock>) {
-        let sighand = current.sighand();
-        let exec_task = if sighand.flags_contains(SignalFlags::GROUP_EXEC) {
-            sighand.group_exec_task()
-        } else {
-            None
-        };
-        let is_mt_exec_leader = current.is_thread_group_leader()
-            && exec_task
-                .as_ref()
-                .map(|t| !Arc::ptr_eq(t, current))
-                .unwrap_or(false);
-        if sighand.flags_contains(SignalFlags::GROUP_EXEC) {
-            if let Some(exec_task) = exec_task.as_ref() {
-                if !Arc::ptr_eq(exec_task, current) {
-                    let notify_count = sighand.group_exec_notify_count();
-                    if notify_count < 0 {
-                        // mt-exec: exec 线程正在等待 leader 退出
-                        sighand.wake_group_exec_waiters();
-                    } else if !current.is_thread_group_leader() {
-                        sighand.dec_group_exec_notify_count_and_wake();
-                    }
-                }
-            }
-            let should_clear = exec_task
-                .as_ref()
-                .map(|t| Arc::ptr_eq(t, current))
-                .unwrap_or(false);
-            if should_clear {
-                sighand.finish_group_exec();
-            }
-        }
-        // mt-exec: leader 退出时只标记 zombie，避免触发普通退出通知/收养
-        if is_mt_exec_leader {
-            current.set_exit_state_zombie();
-            Self::wake_pidfd_pollers_for_task_exit(current);
-            return;
-        }
-        // 让INIT进程收养所有子进程
-        if current.raw_pid() != RawPid(1) {
-            unsafe {
-                current
-                    .adopt_childen()
-                    .unwrap_or_else(|e| panic!("adopte_childen failed: error: {e:?}"))
-            };
-            ProcessManager::exit_ptrace(current);
-            // 在通知父进程之前，先标记为 Zombie，保证 wait 可见
-            current.set_exit_state_zombie();
-            Self::wake_pidfd_pollers_for_task_exit(current);
-            let r = current.parent_pcb.read_irqsave().upgrade();
-            if r.is_none() {
-                return;
-            }
-            let parent_pcb = r.unwrap();
-
-            // 检查子进程的 exit_signal，只有在正信号编号时才发送信号。
-            // Linux 语义中 exit_signal=0 表示不发信号但仍可 wait，-1 表示非 leader 线程。
-            let exit_signal = current.exit_signal.load(Ordering::SeqCst);
-            let sigchld_disposition = parent_pcb.sighand().handler(Signal::SIGCHLD);
-            let sigchld_ignored = sigchld_disposition
-                .as_ref()
-                .map(|sa| sa.is_ignore())
-                .unwrap_or(false);
-            let sigchld_no_cldwait = sigchld_disposition
-                .as_ref()
-                .map(|sa| sa.flags().contains(SigFlags::SA_NOCLDWAIT))
-                .unwrap_or(false);
-            let autoreap = !current.is_ptraced()
-                && exit_signal == Signal::SIGCHLD as i32
-                && (sigchld_ignored || sigchld_no_cldwait);
-            let is_kthread = current.is_kthread();
-            if autoreap && current.try_mark_dead_from_zombie() {
-                unsafe { ProcessManager::release(current.raw_pid()) };
-            }
-
-            if exit_signal > 0 && !(autoreap && sigchld_ignored) {
-                let r = crate::ipc::kill::send_signal_to_pcb(
-                    parent_pcb.clone(),
-                    Signal::from(exit_signal),
-                );
-                if let Err(e) = r {
-                    warn!(
-                        "failed to send kill signal to {:?}'s parent pcb {:?}: {:?}",
-                        current.raw_pid(),
-                        parent_pcb.raw_pid(),
-                        e
-                    );
-                }
-            }
-
-            // Wake the wait parent regardless of exit_signal value.
-            // exit_signal only determines which signal to send, not whether to wake wait.
-            // DragonOS waiters sleep on per-task wait_queue, so we must also wake
-            // the parent's thread-group leader to compensate for Linux's shared
-            // signal->wait_chldexit queue semantics.
-            ProcessManager::wake_wait_parent(&parent_pcb);
-
-            // kthread 退出时显式唤醒 kthreadd，使其回收 zombie
-            if is_kthread {
-                let _ = ProcessManager::wakeup(&parent_pcb);
-            }
-
-            // todo: 这里还需要根据线程组的信息，决定信号的发送
-        }
-    }
-
-    fn wake_pidfd_pollers_for_task_exit(task: &Arc<ProcessControlBlock>) {
-        let thread_pid = task.pid();
-        thread_pid.wake_pidfd_pollers();
-        if let Some(tgid_pid) = task.task_pid_ptr(PidType::TGID) {
-            if !Arc::ptr_eq(&thread_pid, &tgid_pid) {
-                tgid_pid.wake_pidfd_pollers();
-            }
-        }
-    }
-
-    /// 退出当前进程
-    ///
-    /// ## 参数
-    ///
-    /// - `exit_code` : 进程的退出码
-    ///
-    /// ## 注意
-    ///  对于正常退出的进程，状态码应该先左移八位，以便用户态读取的时候正常返回退出码；而对于被信号终止的进程，状态码则是最低七位，无需进行移位操作。
-    ///
-    ///  因此注意，传入的`exit_code`应该是已经完成了移位操作的
-    pub fn exit(exit_code: usize) -> ! {
-        // 检查是否是init进程尝试退出，如果是则产生panic
-        let current_pcb = ProcessManager::current_pcb();
-
-        if current_pcb.raw_pid() == RawPid(0) {
-            log::error!(
-                "Idle process (pid=0) attempted to exit with code {}. Halting current cpu.",
-                exit_code
-            );
-            loop {
-                spin_loop();
-            }
-        }
-
-        if current_pcb.raw_pid() == RawPid(1) {
-            log::error!(
-                "Init process (pid=1) attempted to exit with code {}. This should not happen and indicates a serious system error.",
-                exit_code
-            );
-            loop {
-                spin_loop();
-            }
-        }
-
-        let pid: Arc<Pid>;
-        let raw_pid = current_pcb.raw_pid();
-        // log::debug!("[exit: {}]", raw_pid.data());
-        {
-            let pcb = current_pcb.clone();
-            pcb.mark_exiting();
-            pid = pcb.pid();
-            if pid.is_child_reaper() {
-                pid.ns_of_pid().disable_pid_allocation();
-            }
-            pcb.wait_queue.mark_dead();
-
-            // 进行进程退出后的工作
-            let thread = pcb.thread.write_irqsave();
-            let clear_child_tid = thread.clear_child_tid;
-            let vfork_done = thread.vfork_done.clone();
-            drop(thread);
-            if let Some(addr) = clear_child_tid {
-                // 按 Linux 语义：先清零 userland 的 *clear_child_tid，再 futex_wake(addr)
-                let cleared_ok = unsafe {
-                    match clear_user_protected(addr, core::mem::size_of::<i32>()) {
-                        Ok(_) => true,
-                        Err(e) => {
-                            // clear_child_tid 指针可能无效或已拆映射：不能因此 panic。
-                            warn!("clear tid failed: {e:?}");
-                            false
-                        }
-                    }
-                };
-                // 若无法清零 *clear_child_tid，则也不要 futex_wake（避免进一步的非法用户态访问）
-                if cleared_ok
-                    && Arc::strong_count(&pcb.basic().user_vm().expect("User VM Not found")) > 1
-                {
-                    // Linux 使用 FUTEX_SHARED 标志来唤醒 clear_child_tid
-                    // 这允许跨进程/线程的同步（例如 pthread_join）
-                    let _ =
-                        Futex::futex_wake(addr, FutexFlag::FLAGS_SHARED, 1, FUTEX_BITSET_MATCH_ANY);
-                }
-            }
-            compiler_fence(Ordering::SeqCst);
-
-            RobustListHead::exit_robust_list(pcb.clone());
-            // 如果是vfork出来的进程，则需要处理completion
-            if let Some(vd) = vfork_done {
-                vd.complete_all();
-            }
-
-            // Linux exit_mm() happens before exit_files(): after clear_child_tid
-            // and robust-list cleanup have consumed user memory, the exiting task
-            // must stop exposing a user-visible mm even if file close blocks.
-            //
-            // The CPU-visible active mm is switched to idle here and tracked by
-            // per-CPU TlbState. The PCB-visible user_vm becomes None, matching
-            // Linux task->mm == NULL after exit_mm().
-            let old_user_vm = {
-                let idle_vm = IDLE_PROCESS_ADDRESS_SPACE();
-                let irq_guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-                let cpu = smp_get_processor_id();
-                let mut basic = pcb.basic_mut();
-                let old_vm = unsafe { basic.replace_user_vm(None) };
-                idle_vm.active_cpus_set(cpu);
-                unsafe { idle_vm.make_current() };
-                if let Some(old_vm) = old_vm.as_ref() {
-                    old_vm.active_cpus_clear(cpu);
-                }
-                unsafe { crate::mm::tlb::tlb_state_set_loaded_mm(idle_vm) };
-                drop(basic);
-                drop(irq_guard);
-                old_vm
-            };
-            if let Some(old_vm) = old_user_vm.as_ref() {
-                if !Self::mm_has_user_tasks(old_vm) {
-                    unsafe {
-                        old_vm.write().unmap_all();
-                    }
-                }
-            }
-            drop(old_user_vm);
-
-            pcb.exit_files();
-            pcb.exit_timers();
-
-            let (current_tty, is_session_leader, sid) = {
-                let siginfo = pcb.sig_info_irqsave();
-                (siginfo.tty(), siginfo.is_session_leader, pcb.task_session())
-            };
-            if let Some(tty) = current_tty {
-                if is_session_leader {
-                    let tty_pgrp = tty.core().contorl_info_irqsave().pgid.clone();
-                    if let Some(pgrp) = tty_pgrp {
-                        let _ = crate::ipc::kill::send_signal_to_pgid(&pgrp, Signal::SIGHUP);
-                    }
-                    TtyJobCtrlManager::remove_session_tty(&tty);
-                    if let Some(sid) = sid {
-                        TtyJobCtrlManager::session_clear_tty(sid);
-                    } else {
-                        pcb.sig_info_mut().set_tty(None);
-                    }
-                } else {
-                    let mut g = tty.core().contorl_info_irqsave();
-                    if g.pgid == Some(pid) {
-                        g.pgid = None;
-                    }
-                    drop(g);
-                    pcb.sig_info_mut().set_tty(None);
-                }
-            } else {
-                pcb.sig_info_mut().set_tty(None);
-            }
-
-            // Linux 语义：zombie 不应出现在 cgroup.procs 中。
-            // 必须持有 cgroup_accounting_lock 以避免与 cgroup.procs 写入死锁
-            {
-                let _cgroup_guard = crate::cgroup::cgroup_accounting_lock().lock();
-                pcb.task_cgroup_node().remove_task(raw_pid);
-            }
-            if pcb.is_kthread() {
-                let exited_completion = {
-                    let worker_private = pcb.worker_private();
-                    worker_private
-                        .as_ref()
-                        .and_then(|x| x.kernel_thread())
-                        .map(|x| x.exited_completion())
-                };
-                if let Some(exited_completion) = exited_completion {
-                    exited_completion.complete_all();
-                }
-            }
-
-            // 对标 Linux do_task_dead()：不在此处手动 deactivate_task。
-            // 任务仍留在 rq 上，由后续 __schedule() 检测到 Exited 状态后
-            // 执行唯一的 deactivate_task，避免双重 dequeue 导致 nr_running 下溢。
-
-            let _final_irq_guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-            // 在最后，调用 exit_notify 之前，设置调度状态为 Exited
-            // 对标 Linux do_task_dead() 中 set_special_state(TASK_DEAD)：
-            // 在 pi_lock 保护下完成状态写入，序列化与 wakeup() 中 pi_lock 保护的并发唤醒。
-            {
-                let _pi_guard = pcb.sched_info.pi_lock_irqsave();
-                pcb.sched_info.set_state(ProcessState::Exited(exit_code));
-            }
-            ProcessManager::exit_notify(&pcb);
-
-            __schedule_with_current(SchedMode::SM_NONE, current_pcb);
-            error!("raw_pid {raw_pid:?} exited but sched again!");
-            #[allow(clippy::empty_loop)]
-            loop {
-                spin_loop();
-            }
-        }
-    }
-
-    /// 线程组整体退出（仿照 Linux do_group_exit 语义）
-    ///
-    /// - `exit_code`:
-    ///   - 对于 sys_exit_group：应当已经左移 8 位（wstatus 编码）
-    ///   - 对于致命信号：低 7 位为信号编号，无需移位
-    pub fn group_exit(exit_code: usize) -> ! {
-        let final_exit_code: usize;
-
-        // 使用闭包，确保这里所有指针都释放了
-        {
-            // 检查是否是 init 进程
-            let current_pcb = ProcessManager::current_pcb();
-
-            if current_pcb.raw_pid() == RawPid(1) {
-                log::error!(
-                "Init process (pid=1) attempted to group_exit with code {}. This should not happen and indicates a serious system error.",
-                exit_code
-            );
-                loop {
-                    spin_loop();
-                }
-            }
-
-            // 1. 在共享的 sighand 上原子性地设置 GROUP_EXIT 标志与 group_exit_code
-            //    若已有线程抢先设置，则复用已存在的退出码。
-            let sighand = current_pcb.sighand();
-            current_pcb.mark_exiting();
-            final_exit_code = sighand.start_group_exit(exit_code);
-
-            // 2. 向同一线程组内的其他线程发送 SIGKILL，唤醒并强制终止它们
-            //    仿照 Linux zap_other_threads 的语义，这里仅负责投递 SIGKILL，
-            //    实际退出在各线程上下文中完成。
-            {
-                let send_sigkill_thread = |task: Arc<ProcessControlBlock>| {
-                    if task.flags().contains(ProcessFlags::EXITING) {
-                        return;
-                    }
-                    let mut info = SigInfo::new(
-                        Signal::SIGKILL,
-                        0,
-                        SigCode::Kernel,
-                        SigType::Kill {
-                            pid: RawPid::new(0),
-                            uid: 0,
-                        },
-                    );
-                    let _ = Signal::SIGKILL.send_signal_info_to_pcb(
-                        Some(&mut info),
-                        task,
-                        PidType::PID,
-                    );
-                };
-
-                // 统一从线程组组长的 ThreadInfo 中获取完整的线程列表，
-                // 避免从非组长线程看到的 group_tasks 为空导致遗漏。
-                let leader = {
-                    let ti = current_pcb.thread.read_irqsave();
-                    ti.group_leader().unwrap_or_else(|| current_pcb.clone())
-                };
-
-                let group_tasks = {
-                    let ti = leader.threads_read_irqsave();
-                    ti.group_tasks.clone()
-                };
-
-                // 先尝试对组长发送 SIGKILL（若当前线程不是组长）
-                if !Arc::ptr_eq(&leader, &current_pcb)
-                    && !leader.flags().contains(ProcessFlags::EXITING)
-                {
-                    send_sigkill_thread(leader.clone());
-                }
-
-                // 再遍历组长维护的 group_tasks，向其他线程发送 SIGKILL
-                for weak in group_tasks {
-                    if let Some(task) = weak.upgrade() {
-                        // 跳过当前线程自身
-                        if task.raw_pid() == current_pcb.raw_pid() {
-                            continue;
-                        }
-                        if task.flags().contains(ProcessFlags::EXITING) {
-                            continue;
-                        }
-                        send_sigkill_thread(task);
-                    }
-                }
-            }
-
-            drop(current_pcb);
-        }
-        // 3. 当前线程按照统一的退出码执行常规退出流程
-
-        ProcessManager::exit(final_exit_code);
-    }
-
-    /// 从全局进程列表中删除一个进程，并从父进程的 children 列表中移除
-    ///
-    /// # 参数
-    ///
-    /// - `pid` : 进程的**全局** pid
-    ///
-    /// # 注意
-    ///
-    /// 调用此函数前，调用者**不能持有**父进程的 children 锁，否则会死锁。
-    pub(super) unsafe fn release(pid: RawPid) {
-        let pcb = ProcessManager::find(pid);
-        if let Some(ref pcb) = pcb {
-            ProcessManager::exit_ptrace(pcb);
-            ProcessManager::ptrace_unlink_tracee(pcb);
-
-            let parent_child_vpid = pcb.real_parent_pcb().and_then(|parent| {
-                let parent_ns = parent.active_pid_ns();
-                pcb.task_pid_nr_ns(PidType::PID, Some(parent_ns))
-                    .map(|vpid| (parent, vpid))
-            });
-
-            pcb.__exit_signal();
-            {
-                let _cgroup_guard = crate::cgroup::cgroup_accounting_lock().lock();
-                pcb.task_cgroup_node().uncharge_pids(1);
-            }
-            // 从父进程的 children 列表中移除
-            if let Some((parent, vpid)) = parent_child_vpid {
-                let mut children = parent.children.write();
-                children.retain(|&p| p != vpid);
-            }
-
-            ALL_PROCESS.lock_irqsave().as_mut().unwrap().remove(&pid);
-        }
-    }
-
-    pub fn ptrace_unlink_tracee(tracee: &Arc<ProcessControlBlock>) {
-        ptrace::unlink_tracee(tracee)
-    }
-
-    pub fn exit_ptrace(tracer: &Arc<ProcessControlBlock>) {
-        ptrace::exit_ptrace(tracer)
-    }
-
-    /// Wake waiters that can observe child state changes through `parent`.
-    ///
-    /// Linux uses the thread-group shared `signal->wait_chldexit` queue.
-    /// DragonOS currently has one `wait_queue` per task, while `do_wait()`
-    /// sleeps on the caller's thread-group leader. Any path that publishes a
-    /// wait-visible transition through a parent relation must therefore wake
-    /// both the concrete parent task and its thread-group leader.
-    pub(crate) fn wake_wait_parent(parent: &Arc<ProcessControlBlock>) {
-        parent
-            .wait_queue
-            .wakeup_all(Some(ProcessState::Blocked(true)));
-
-        let parent_group_leader = {
-            let ti = parent.thread.read_irqsave();
-            ti.group_leader()
-        };
-        if let Some(leader) = parent_group_leader {
-            if !Arc::ptr_eq(&leader, parent) {
-                leader
-                    .wait_queue
-                    .wakeup_all(Some(ProcessState::Blocked(true)));
-            }
-        }
-    }
-
-    /// 上下文切换完成后的钩子函数
-    unsafe fn switch_finish_hook() {
-        // debug!("switch_finish_hook");
-        let prev_pcb = PROCESS_SWITCH_RESULT
-            .as_mut()
-            .unwrap()
-            .get_mut()
-            .prev_pcb
-            .take()
-            .expect("prev_pcb is None");
-        let next_pcb = PROCESS_SWITCH_RESULT
-            .as_mut()
-            .unwrap()
-            .get_mut()
-            .next_pcb
-            .take()
-            .expect("next_pcb is None");
-
-        // 由于进程切换前使用了SpinLockGuard::leak()，所以这里需要手动释放锁
-        fence(Ordering::SeqCst);
-
-        prev_pcb.arch_info.force_unlock();
-        fence(Ordering::SeqCst);
-
-        next_pcb.arch_info.force_unlock();
-        fence(Ordering::SeqCst);
-
-        let migrate_prev_to = PROCESS_SWITCH_RESULT
-            .as_mut()
-            .unwrap()
-            .get_mut()
-            .migrate_prev_to
-            .take();
-
-        if let Some(dest_cpu) = migrate_prev_to {
-            debug_assert!(!Arc::ptr_eq(&prev_pcb, &next_pcb));
-            prev_pcb.sched_info().set_on_cpu(None);
-            enqueue_task_on_cpu(&prev_pcb, dest_cpu, WakeupFlags::WF_MIGRATED, false);
-        }
-
-        let set_child_tid = next_pcb.thread.write_irqsave().set_child_tid.take();
-        if let Some(addr) = set_child_tid {
-            // 对齐 Linux schedule_tail 语义：子任务首次运行时 best-effort 写入 tid，
-            // 失败不影响线程继续执行。
-            let child_tid = next_pcb.task_pid_vnr().data() as i32;
-            let _ = unsafe { write_one_to_user_protected(addr, &child_tid) };
-        }
-    }
-
-    /// 如果目标进程正在目标CPU上运行，那么就让这个cpu陷入内核态
-    ///
-    /// ## 参数
-    ///
-    /// - `pcb` : 进程的pcb
-    #[allow(dead_code)]
-    pub fn kick(pcb: &Arc<ProcessControlBlock>) {
-        ProcessManager::current_pcb().preempt_disable();
-        let cpu_id = pcb.sched_info().on_cpu();
-
-        if let Some(cpu_id) = cpu_id {
-            let current_cpu_id = smp_get_processor_id();
-            // DragonOS 目前没有 Linux `kick_process()` 那样的 lockless rq.current 约定，
-            // 远端 rq.current 必须在持目标 rq 锁时读取。
-            let should_kick = if cpu_id != current_cpu_id {
-                let rq = cpu_rq(cpu_id.data() as usize);
-                let (rq, _guard) = rq.self_lock();
-                Arc::ptr_eq(&rq.current(), pcb)
-            } else {
-                false
-            };
-
-            // Do not kick the current CPU, as it is already running and cannot preempt itself.
-            if should_kick {
-                kick_cpu(cpu_id).expect("ProcessManager::kick(): Failed to kick cpu");
-            }
-        }
-
-        ProcessManager::current_pcb().preempt_enable();
-    }
-}
-
-/// 上下文切换的钩子函数,当这个函数return的时候,将会发生上下文切换
+/// Context switch hook function. When this function returns, a context switch
+/// will occur.
 #[cfg(target_arch = "x86_64")]
 #[inline(never)]
 pub unsafe extern "sysv64" fn switch_finish_hook() {
@@ -1394,2577 +80,4 @@ pub unsafe extern "sysv64" fn switch_finish_hook() {
 #[inline(always)]
 pub unsafe fn switch_finish_hook() {
     ProcessManager::switch_finish_hook();
-}
-
-int_like!(RawPid, AtomicRawPid, usize, AtomicUsize);
-
-impl fmt::Display for RawPid {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.0)
-    }
-}
-
-impl FromStr for RawPid {
-    type Err = core::num::ParseIntError;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        let pid = usize::from_str(s)?;
-        Ok(RawPid(pid))
-    }
-}
-
-impl RawPid {
-    /// 该RawPid暂未分配，待会会初始化它。
-    /// 这个状态只应当出现在进程/线程创建的过程中
-    pub const UNASSIGNED: RawPid = RawPid(usize::MAX - 1);
-    pub const MAX_VALID: RawPid = RawPid(usize::MAX - 32);
-
-    pub fn is_valid(&self) -> bool {
-        self.0 >= Self::MAX_VALID.0
-    }
-}
-
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum ProcessState {
-    /// The process is running on a CPU or in a run queue.
-    Runnable,
-    /// The process is waiting for an event to occur.
-    /// 其中的bool表示该等待过程是否可以被打断。
-    /// - 如果该bool为true,那么，硬件中断/信号/其他系统事件都可以打断该等待过程，使得该进程重新进入Runnable状态。
-    /// - 如果该bool为false,那么，这个进程必须被显式的唤醒，才能重新进入Runnable状态。
-    Blocked(bool),
-    /// 进程被信号终止
-    Stopped,
-    /// The process has exited; usize holds the raw wait status used by Linux wait(2) family.
-    ///
-    /// Normal exit: `(exit_code & 0xff) << 8`; signal termination: signal number in the low 7 bits.
-    /// wait4/waitpid return this value as-is; only waitid `si_status` needs decoding from it.
-    Exited(usize),
-}
-
-#[repr(u8)]
-#[derive(Clone, Copy, Debug, Eq, PartialEq)]
-pub enum ExitState {
-    Running = 0,
-    Zombie = 1,
-    Dead = 2,
-}
-
-impl ExitState {
-    fn from_u8(v: u8) -> Self {
-        match v {
-            1 => ExitState::Zombie,
-            2 => ExitState::Dead,
-            _ => ExitState::Running,
-        }
-    }
-}
-
-mod state_bits {
-    pub const TASK_RUNNING: u32 = 0x0000;
-    pub const TASK_INTERRUPTIBLE: u32 = 0x0001;
-    pub const TASK_UNINTERRUPTIBLE: u32 = 0x0002;
-    pub const TASK_STOPPED: u32 = 0x0004;
-    pub const TASK_DEAD_MARKER: u32 = 0x0080;
-    pub const EXIT_CODE_SHIFT: u32 = 12;
-}
-
-#[allow(dead_code)]
-impl ProcessState {
-    #[inline(always)]
-    pub fn is_runnable(&self) -> bool {
-        return matches!(self, ProcessState::Runnable);
-    }
-
-    #[inline(always)]
-    pub fn is_blocked(&self) -> bool {
-        return matches!(self, ProcessState::Blocked(_));
-    }
-
-    #[inline(always)]
-    pub fn is_blocked_interruptable(&self) -> bool {
-        return matches!(self, ProcessState::Blocked(true));
-    }
-
-    /// Returns `true` if the process state is [`Exited`].
-    #[inline(always)]
-    pub fn is_exited(&self) -> bool {
-        return matches!(self, ProcessState::Exited(_));
-    }
-
-    /// Returns `true` if the process state is [`Stopped`].
-    ///
-    /// [`Stopped`]: ProcessState::Stopped
-    #[inline(always)]
-    pub fn is_stopped(&self) -> bool {
-        matches!(self, ProcessState::Stopped)
-    }
-
-    /// Returns raw wait status if the process state is [`Exited`].
-    #[inline(always)]
-    pub fn raw_wstatus(&self) -> Option<usize> {
-        match self {
-            ProcessState::Exited(code) => Some(*code),
-            _ => None,
-        }
-    }
-
-    /// Returns raw wait status if the process state is [`Exited`].
-    ///
-    /// Kept for existing call sites; new wait code should prefer
-    /// [`ProcessState::raw_wstatus`] to avoid confusing raw wait status with
-    /// the user-visible exit code.
-    #[inline(always)]
-    pub fn exit_code(&self) -> Option<usize> {
-        self.raw_wstatus()
-    }
-
-    #[inline]
-    pub fn to_u32(self) -> u32 {
-        match self {
-            ProcessState::Runnable => state_bits::TASK_RUNNING,
-            ProcessState::Blocked(true) => state_bits::TASK_INTERRUPTIBLE,
-            ProcessState::Blocked(false) => state_bits::TASK_UNINTERRUPTIBLE,
-            ProcessState::Stopped => state_bits::TASK_STOPPED,
-            ProcessState::Exited(code) => {
-                state_bits::TASK_DEAD_MARKER | ((code as u32) << state_bits::EXIT_CODE_SHIFT)
-            }
-        }
-    }
-
-    #[inline]
-    pub fn from_u32(val: u32) -> Self {
-        if val & state_bits::TASK_DEAD_MARKER != 0 {
-            let code = (val >> state_bits::EXIT_CODE_SHIFT) as usize;
-            ProcessState::Exited(code)
-        } else {
-            match val {
-                v if v == state_bits::TASK_RUNNING => ProcessState::Runnable,
-                v if v == state_bits::TASK_INTERRUPTIBLE => ProcessState::Blocked(true),
-                v if v == state_bits::TASK_UNINTERRUPTIBLE => ProcessState::Blocked(false),
-                v if v == state_bits::TASK_STOPPED => ProcessState::Stopped,
-                _ => {
-                    panic!(
-                        "ProcessState::from_u32: corrupted state value 0x{val:08x}, \
-                         this indicates memory corruption or a kernel bug"
-                    );
-                }
-            }
-        }
-    }
-}
-
-bitflags! {
-    /// pcb的标志位
-    pub struct ProcessFlags: usize {
-        /// 当前pcb表示一个内核线程
-        const KTHREAD = 1 << 0;
-        /// 当前进程需要被调度
-        const NEED_SCHEDULE = 1 << 1;
-        /// 进程由于vfork而与父进程存在资源共享
-        const VFORK = 1 << 2;
-        /// 进程不可被冻结
-        const NOFREEZE = 1 << 3;
-        /// 进程正在退出
-        const EXITING = 1 << 4;
-        /// 进程由于接收到终止信号唤醒
-        const WAKEKILL = 1 << 5;
-        /// 进程由于接收到信号而退出.(Killed by a signal)
-        const SIGNALED = 1 << 6;
-        /// 进程需要迁移到其他cpu上
-        const NEED_MIGRATE = 1 << 7;
-        /// 随机化的虚拟地址空间，主要用于动态链接器的加载
-        const RANDOMIZE = 1 << 8;
-        /// 进程有未处理的信号（这是一个用于快速判断的标志位）
-        /// 相当于Linux的TIF_SIGPENDING
-        const HAS_PENDING_SIGNAL = 1 << 9;
-        /// 进程需要恢复之前保存的信号掩码
-        const RESTORE_SIG_MASK = 1 << 10;
-        /// Forked but didn't exec
-        const FORKNOEXEC = 1 << 11;
-        /// 进程需要在返回用户态前处理 rseq
-        const NEED_RSEQ = 1 << 12;
-        /// 进程正在等待 IO 操作完成（用于 iowait 统计）
-        const IN_IOWAIT = 1 << 13;
-        /// 线程组 exec 期间延迟 PID/TGID/PGID/SID 的 unhash
-        const DEFER_UNHASH = 1 << 14;
-        /// PID links and visible-thread accounting have already been released.
-        const PID_UNHASHED = 1 << 15;
-        /// Task is currently traced by another task.
-        const PTRACED = 1 << 16;
-    }
-}
-
-impl ProcessFlags {
-    pub const fn fork_inherited(&self) -> Self {
-        Self::from_bits_truncate(self.bits & Self::RANDOMIZE.bits)
-    }
-
-    pub const fn exit_to_user_mode_work(&self) -> Self {
-        Self::from_bits_truncate(
-            self.bits
-                & (Self::NEED_SCHEDULE.bits | Self::HAS_PENDING_SIGNAL.bits | Self::NEED_RSEQ.bits),
-        )
-    }
-
-    /// 测试并清除标志位
-    ///
-    /// ## 参数
-    ///
-    /// - `rhs` : 需要测试并清除的标志位
-    ///
-    /// ## 返回值
-    ///
-    /// 如果标志位在清除前是置位的，则返回 `true`，否则返回 `false`
-    pub const fn test_and_clear(&mut self, rhs: Self) -> bool {
-        let r = (self.bits & rhs.bits) != 0;
-        self.bits &= !rhs.bits;
-        r
-    }
-}
-
-#[derive(Debug, Default)]
-pub struct CpuItimer {
-    pub value: u64,    // 剩余时间 ns
-    pub interval: u64, // 间隔时间 ns
-    pub is_active: bool,
-}
-
-#[derive(Debug, Clone)]
-pub struct ProcessItimer {
-    pub timer: Arc<crate::time::timer::Timer>,
-    pub config: crate::time::syscall::Itimerval,
-}
-
-#[derive(Debug, Default)]
-pub struct ProcessItimers {
-    pub real: Option<ProcessItimer>, // 用于 ITIMER_REAL
-    pub virt: CpuItimer,             // 用于 ITIMER_REAL
-    pub prof: CpuItimer,             // 用于 ITIMER_PROF
-}
-
-#[derive(Debug)]
-pub struct ProcessControlBlock {
-    /// 当前进程的pid
-    pid: AtomicRawPid,
-    /// 当前进程的线程组id（这个值在同一个线程组内永远不变）
-    tgid: RawPid,
-
-    thread_pid: RwLock<Option<Arc<Pid>>>,
-    /// PID链接数组
-    pid_links: [PidLink; PidType::PIDTYPE_MAX],
-
-    /// namespace代理
-    nsproxy: RcuArcSlot<NsProxy>,
-    /// 任务所属 cgroup（v2）
-    task_cgroup: RwLock<TaskCgroupRef>,
-
-    basic: RwLock<ProcessBasicInfo>,
-    /// 当前进程的自旋锁持有计数
-    preempt_count: AtomicUsize,
-    /// 当前进程的 RCU 读侧嵌套层数
-    rcu_read_depth: AtomicUsize,
-
-    flags: LockFreeFlags<ProcessFlags>,
-    /// 当前任务是否已经计入全局可见线程数。
-    visible_thread_accounted: AtomicBool,
-    /// Serializes task-local pointer publication for RCU-protected metadata
-    /// such as `cred`, `nsproxy`, and `sighand`.
-    task_lock: SpinLock<()>,
-    worker_private: SpinLock<Option<WorkerPrivate>>,
-    /// 进程的内核栈
-    kernel_stack: RwLock<KernelStack>,
-
-    /// 系统调用栈
-    syscall_stack: RwLock<KernelStack>,
-
-    /// 与调度相关的信息
-    sched_info: ProcessSchedulerInfo,
-    /// 与处理器架构相关的信息
-    arch_info: SpinLock<ArchPCBInfo>,
-    /// 与信号处理相关的信息(似乎可以是无锁的)
-    sig_info: RwLock<ProcessSignalInfo>,
-    sighand: RcuArcSlot<SigHand>,
-    /// 备用信号栈
-    sig_altstack: RwLock<SigStackArch>,
-    /// 退出状态（Running/Zombie/Dead）
-    exit_state: AtomicU8,
-
-    /// Linux task_struct::exit_signal 语义：
-    /// - -1: 非线程组 leader（CLONE_THREAD）；
-    /// - 0: 不发退出信号，但仍是可等待的 clone 子进程；
-    /// - >0: 退出时通知父进程的信号编号。
-    exit_signal: AtomicI32,
-    /// 父进程退出时要发送给当前进程的信号（PR_SET_PDEATHSIG）
-    pdeath_signal: AtomicSignal,
-
-    /// prctl(PR_SET/GET_NO_NEW_PRIVS) 状态：线程级（task）语义。
-    no_new_privs: AtomicBool,
-
-    /// prctl(PR_SET/GET_KEEPCAPS) 状态：线程级（task）语义。
-    /// 当为 true 时，进程改变 UID/GID 后会保留 capabilities。
-    keepcaps: AtomicBool,
-
-    /// prctl(PR_SET/GET_DUMPABLE) 状态。
-    /// Linux: 0=SUID_DUMP_DISABLE, 1=SUID_DUMP_USER；2(SUID_DUMP_ROOT) 不允许通过 PR_SET_DUMPABLE 设置。
-    dumpable: AtomicU8,
-
-    pub(crate) seccomp_mode: AtomicU8,
-    pub(crate) seccomp_filter: SpinLock<Option<Arc<seccomp::SeccompFilter>>>,
-
-    /// 父进程指针
-    pub(crate) parent_pcb: RwLock<Weak<ProcessControlBlock>>,
-    /// 真实父进程指针
-    pub(crate) real_parent_pcb: RwLock<Weak<ProcessControlBlock>>,
-    /// wait 自然父进程指针。
-    ///
-    /// Linux wait_task_zombie()/wait_consider_task() 中 __WNOTHREAD 依赖
-    /// task_struct::parent，而 DragonOS 的 parent_pcb/real_parent_pcb 按线程组
-    /// leader 建模。该字段保留 wait 所需的线程级父关系。
-    pub(crate) wait_parent_pcb: RwLock<Weak<ProcessControlBlock>>,
-    /// Thread-level natural-parent compensation for PTRACE_TRACEME.
-    ///
-    /// On normal fork it points to the creating thread, compensating for the
-    /// fact that real_parent_pcb is still modeled on the thread-group leader.
-    /// After CLONE_PARENT, reparent, and de_thread it must be updated to follow
-    /// Linux real_parent semantics. It is NOT an immutable “first-fork creator”
-    /// record.
-    pub(crate) fork_parent_pcb: RwLock<Weak<ProcessControlBlock>>,
-
-    /// 子进程链表
-    children: RwLock<Vec<RawPid>>,
-    /// Tasks currently traced by this process. Entries are global raw pids.
-    ptraced: RwLock<Vec<RawPid>>,
-    /// Current tracer if this process is ptraced.
-    ptracer_pcb: RwLock<Weak<ProcessControlBlock>>,
-
-    /// 等待队列
-    wait_queue: WaitQueue,
-
-    /// CPU-time 等待队列：用于 CLOCK_{PROCESS,THREAD}_CPUTIME_ID 的 clock_nanosleep
-    cputime_wait_queue: WaitQueue,
-
-    /// 线程信息
-    thread: RwLock<ThreadInfo>,
-
-    /// 进程文件系统的状态
-    fs: RwLock<Arc<FsStruct>>,
-
-    ///闹钟定时器
-    alarm_timer: SpinLock<Option<AlarmTimer>>,
-    itimers: SpinLock<ProcessItimers>,
-    /// POSIX interval timers（timer_create/timer_settime/...）
-    posix_timers: SpinLock<posix_timer::ProcessPosixTimers>,
-
-    /// CPU时间片
-    cpu_time: Arc<ProcessCpuTime>,
-    /// 已退出线程的线程组级资源累计。对齐 Linux signal_struct 的已退出线程统计。
-    exited_thread_group_rusage: SpinLock<RUsage>,
-    /// 已被 wait 系列成功回收的子进程资源累计，对齐 getrusage(RUSAGE_CHILDREN)。
-    children_rusage: SpinLock<RUsage>,
-
-    /// 进程的robust lock列表
-    robust_list: RwLock<Option<RobustListHead>>,
-
-    /// rseq（Restartable Sequences）状态
-    rseq_state: RwLock<rseq::RseqState>,
-
-    /// 进程作为主体的凭证集
-    cred: RcuArcSlot<Cred>,
-    self_ref: Weak<ProcessControlBlock>,
-
-    restart_block: SpinLock<Option<RestartBlock>>,
-
-    /// 进程的可执行文件路径
-    executable_path: RwLock<String>,
-    /// 进程的命令行（用于 /proc/<pid>/cmdline，Linux 语义：argv 以 '\0' 分隔）
-    cmdline: RwLock<Vec<u8>>,
-    /// 资源限制（rlimit）数组
-    rlimits: RwLock<[RLimit64; RLimitID::Nlimits as usize]>,
-}
-
-impl ProcessControlBlock {
-    /// Generate a new pcb.
-    ///
-    /// ## 参数
-    ///
-    /// - `name` : 进程的名字
-    /// - `kstack` : 进程的内核栈
-    ///
-    /// ## 返回值
-    ///
-    /// 返回一个新的pcb
-    pub fn new(name: String, kstack: KernelStack) -> Arc<Self> {
-        return Self::do_create_pcb(name, kstack, false);
-    }
-
-    /// 创建一个新的idle进程
-    ///
-    /// 请注意，这个函数只能在进程管理初始化的时候调用。
-    pub fn new_idle(cpu_id: u32, kstack: KernelStack) -> Arc<Self> {
-        let name = format!("idle-{}", cpu_id);
-        return Self::do_create_pcb(name, kstack, true);
-    }
-
-    /// # 函数的功能
-    ///
-    /// 返回此函数是否是内核进程
-    ///
-    /// # 返回值
-    ///
-    /// 若进程是内核进程则返回true 否则返回false
-    pub fn is_kthread(&self) -> bool {
-        self.flags().contains(ProcessFlags::KTHREAD)
-    }
-
-    #[inline(never)]
-    fn do_create_pcb(name: String, kstack: KernelStack, is_idle: bool) -> Arc<Self> {
-        // 初始化namespace代理
-        let nsproxy = if is_idle {
-            // idle进程使用root namespace
-            NsProxy::new_root()
-        } else {
-            // 其他进程继承父进程的namespace
-            ProcessManager::current_pcb().nsproxy().clone()
-        };
-        let task_cgroup = if is_idle {
-            TaskCgroupRef::new(cgroup_root_node())
-        } else {
-            ProcessManager::current_pcb().task_cgroup_ref()
-        };
-
-        let (raw_pid, ppid, cwd, cred, tty): (
-            RawPid,
-            RawPid,
-            String,
-            Arc<Cred>,
-            Option<Arc<TtyCore>>,
-        ) = if is_idle {
-            let cred = INIT_CRED.clone();
-            (RawPid(0), RawPid(0), "/".to_string(), cred, None)
-        } else {
-            let ppid = ProcessManager::current_pcb().task_pid_vnr();
-            let cred = ProcessManager::current_pcb().cred();
-
-            let cwd = ProcessManager::current_pcb().basic().cwd();
-            let tty = ProcessManager::current_pcb().sig_info_irqsave().tty();
-
-            // Here, UNASSIGNED is used to represent an unallocated pid,
-            // which will be allocated later in `copy_process`.
-            let raw_pid = RawPid::UNASSIGNED;
-
-            (raw_pid, ppid, cwd, cred, tty)
-        };
-
-        let basic_info = ProcessBasicInfo::new(ppid, name.clone(), cwd, None);
-        let preempt_count = AtomicUsize::new(0);
-        let rcu_read_depth = AtomicUsize::new(0);
-        let flags = unsafe { LockFreeFlags::new(ProcessFlags::empty()) };
-        let initial_sighand = SigHand::new();
-        initial_sighand.attach_task_ref();
-
-        let sched_info = ProcessSchedulerInfo::new(None);
-
-        let ppcb: Weak<ProcessControlBlock> = ProcessManager::find_task_by_vpid(ppid)
-            .map(|p| Arc::downgrade(&p))
-            .unwrap_or_default();
-
-        // 使用 Arc::new_cyclic 避免在栈上创建巨大的结构体
-        let pcb = Arc::new_cyclic(|weak| {
-            let arch_info = SpinLock::new(ArchPCBInfo::new(&kstack));
-
-            let pcb = Self {
-                pid: AtomicRawPid::new(raw_pid),
-                tgid: raw_pid,
-                thread_pid: RwLock::new(None),
-                pid_links: core::array::from_fn(|_| PidLink::default()),
-                nsproxy: RcuArcSlot::new(nsproxy),
-                task_cgroup: RwLock::new(task_cgroup),
-                basic: basic_info,
-                preempt_count,
-                rcu_read_depth,
-                flags,
-                visible_thread_accounted: AtomicBool::new(false),
-                task_lock: SpinLock::new(()),
-                kernel_stack: RwLock::new(kstack),
-                syscall_stack: RwLock::new(KernelStack::new().unwrap()),
-                worker_private: SpinLock::new(None),
-                sched_info,
-                arch_info,
-                sig_info: RwLock::new(ProcessSignalInfo::default()),
-                sighand: RcuArcSlot::new(initial_sighand.clone()),
-                sig_altstack: RwLock::new(SigStackArch::new()),
-                exit_state: AtomicU8::new(ExitState::Running as u8),
-                exit_signal: AtomicI32::new(Signal::SIGCHLD as i32),
-                pdeath_signal: AtomicSignal::new(Signal::INVALID),
-
-                no_new_privs: AtomicBool::new(false),
-                keepcaps: AtomicBool::new(false),
-                // 默认设置为 SUID_DUMP_USER(=1)，满足 gVisor 的 SetGetDumpability 预期。
-                dumpable: AtomicU8::new(1),
-                seccomp_mode: AtomicU8::new(seccomp::SeccompMode::Disabled as u8),
-                seccomp_filter: SpinLock::new(None),
-                parent_pcb: RwLock::new(ppcb.clone()),
-                real_parent_pcb: RwLock::new(ppcb.clone()),
-                wait_parent_pcb: RwLock::new(ppcb.clone()),
-                fork_parent_pcb: RwLock::new(ppcb),
-                children: RwLock::new(Vec::new()),
-                ptraced: RwLock::new(Vec::new()),
-                ptracer_pcb: RwLock::new(Weak::new()),
-                wait_queue: WaitQueue::default(),
-                cputime_wait_queue: WaitQueue::default(),
-                thread: RwLock::new(ThreadInfo::new()),
-                fs: RwLock::new(Arc::new(FsStruct::new())),
-                alarm_timer: SpinLock::new(None),
-                itimers: SpinLock::new(ProcessItimers::default()),
-                posix_timers: SpinLock::new(posix_timer::ProcessPosixTimers::default()),
-                cpu_time: Arc::new(ProcessCpuTime::default()),
-                exited_thread_group_rusage: SpinLock::new(RUsage::default()),
-                children_rusage: SpinLock::new(RUsage::default()),
-                robust_list: RwLock::new(None),
-                rseq_state: RwLock::new(rseq::RseqState::new()),
-                cred: RcuArcSlot::new(cred),
-                self_ref: weak.clone(),
-                restart_block: SpinLock::new(None),
-                executable_path: RwLock::new(name),
-                cmdline: RwLock::new(Vec::new()),
-                rlimits: RwLock::new(Self::default_rlimits()),
-            };
-
-            pcb.sig_info.write().set_tty(tty);
-
-            // 初始化系统调用栈
-            #[cfg(target_arch = "x86_64")]
-            pcb.arch_info
-                .lock()
-                .init_syscall_stack(&pcb.syscall_stack.read());
-
-            pcb
-        });
-
-        pcb.sched_info()
-            .sched_entity()
-            .force_mut()
-            .set_pcb(Arc::downgrade(&pcb));
-        // 设置进程的arc指针到内核栈和系统调用栈的最低地址处
-        unsafe {
-            pcb.kernel_stack
-                .write()
-                .set_pcb(Arc::downgrade(&pcb))
-                .unwrap();
-
-            pcb.syscall_stack
-                .write()
-                .set_pcb(Arc::downgrade(&pcb))
-                .unwrap()
-        };
-
-        return pcb;
-    }
-
-    fn default_rlimits() -> [crate::process::resource::RLimit64; RLimitID::Nlimits as usize] {
-        use crate::mm::ucontext::UserStack;
-        use crate::process::resource::{RLimit64, RLimitID};
-
-        let mut arr = [RLimit64 {
-            rlim_cur: 0,
-            rlim_max: 0,
-        }; RLimitID::Nlimits as usize];
-
-        // Linux 典型默认值：软限制1024，硬限制可通过setrlimit调整
-        // 文件描述符表会根据RLIMIT_NOFILE自动扩容
-        arr[RLimitID::Nofile as usize] = RLimit64 {
-            rlim_cur: FileDescriptorVec::MAX_CAPACITY as u64,
-            rlim_max: FileDescriptorVec::MAX_CAPACITY as u64,
-        };
-
-        arr[RLimitID::Stack as usize] = RLimit64 {
-            rlim_cur: UserStack::DEFAULT_USER_STACK_SIZE as u64,
-            rlim_max: UserStack::DEFAULT_USER_STACK_SIZE as u64,
-        };
-
-        arr[RLimitID::As as usize] = {
-            let end = <crate::arch::MMArch as crate::mm::MemoryManagementArch>::USER_END_VADDR;
-            RLimit64 {
-                rlim_cur: end.data() as u64,
-                rlim_max: end.data() as u64,
-            }
-        };
-        arr[RLimitID::Rss as usize] = arr[RLimitID::As as usize];
-
-        // 设置文件大小限制的默认值 (Linux默认通常是unlimited)
-        arr[RLimitID::Fsize as usize] = RLimit64 {
-            rlim_cur: u64::MAX,
-            rlim_max: u64::MAX,
-        };
-
-        // Linux commonly defaults RLIMIT_MEMLOCK to 64 KiB. Keeping the hard
-        // limit non-zero also allows unprivileged tests to lower the soft limit.
-        arr[RLimitID::Memlock as usize] = RLimit64 {
-            rlim_cur: 64 * 1024,
-            rlim_max: 64 * 1024,
-        };
-
-        arr
-    }
-
-    #[inline(always)]
-    pub fn get_rlimit(&self, res: RLimitID) -> crate::process::resource::RLimit64 {
-        self.rlimits.read()[res as usize]
-    }
-
-    pub fn set_rlimit(
-        &self,
-        res: RLimitID,
-        newv: crate::process::resource::RLimit64,
-    ) -> Result<(), system_error::SystemError> {
-        use system_error::SystemError;
-        if newv.rlim_cur > newv.rlim_max {
-            return Err(SystemError::EINVAL);
-        }
-
-        // 注意：允许RLIMIT_NOFILE设置为0，这是测试用例的预期行为
-        // 当rlim_cur为0时，无法分配新的文件描述符，但现有fd仍可使用
-
-        // 对于RLIMIT_NOFILE，检查是否超过系统实现的最大容量限制
-        if res == RLimitID::Nofile {
-            if newv.rlim_cur > FileDescriptorVec::MAX_CAPACITY as u64 {
-                return Err(SystemError::EINVAL);
-            }
-            if newv.rlim_max > FileDescriptorVec::MAX_CAPACITY as u64 {
-                return Err(SystemError::EINVAL);
-            }
-        }
-
-        let cur = self.rlimits.read()[res as usize];
-        if newv.rlim_max > cur.rlim_max {
-            let cred = self.cred();
-            if !cred.has_capability(crate::process::cred::CAPFlags::CAP_SYS_RESOURCE) {
-                return Err(SystemError::EPERM);
-            }
-        }
-
-        // 更新rlimit
-        self.rlimits.write()[res as usize] = newv;
-
-        // 如果是RLIMIT_NOFILE变化，调整文件描述符表
-        if res == RLimitID::Nofile {
-            if let Err(e) = self.adjust_fd_table_for_rlimit_change(newv.rlim_cur as usize) {
-                // 如果调整失败，回滚rlimit设置
-                self.rlimits.write()[res as usize] = cur;
-                return Err(e);
-            }
-        }
-
-        Ok(())
-    }
-
-    /// 继承父进程的全部rlimit
-    pub fn inherit_rlimits_from(&self, parent: &Arc<ProcessControlBlock>) {
-        let src = *parent.rlimits.read();
-        *self.rlimits.write() = src;
-
-        // 继承后调整文件描述符表以匹配新的RLIMIT_NOFILE
-        let nofile_limit = src[RLimitID::Nofile as usize].rlim_cur as usize;
-        if let Err(e) = self.adjust_fd_table_for_rlimit_change(nofile_limit) {
-            // 如果调整失败，记录错误但不影响继承过程
-            error!(
-                "Failed to adjust fd table after inheriting rlimits: {:?}",
-                e
-            );
-        }
-    }
-
-    /// 当RLIMIT_NOFILE变化时调整文件描述符表
-    ///
-    /// ## 参数
-    /// - `new_rlimit_nofile`: 新的RLIMIT_NOFILE值
-    ///
-    /// ## 返回值
-    /// - `Ok(())`: 调整成功
-    /// - `Err(SystemError)`: 调整失败
-    fn adjust_fd_table_for_rlimit_change(
-        &self,
-        new_rlimit_nofile: usize,
-    ) -> Result<(), system_error::SystemError> {
-        let fd_table = self.basic.read().try_fd_table().unwrap();
-        let mut fd_table_guard = fd_table.write();
-        fd_table_guard.adjust_for_rlimit_change(new_rlimit_nofile)
-    }
-
-    /// 返回当前进程的锁持有计数
-    #[inline(always)]
-    pub fn preempt_count(&self) -> usize {
-        return self.preempt_count.load(Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn rcu_read_depth(&self) -> usize {
-        self.rcu_read_depth.load(Ordering::SeqCst)
-    }
-
-    /// 增加当前进程的锁持有计数
-    #[inline(always)]
-    pub fn preempt_disable(&self) {
-        self.preempt_count.fetch_add(1, Ordering::SeqCst);
-    }
-
-    /// 减少当前进程的锁持有计数
-    #[inline(always)]
-    pub fn preempt_enable(&self) {
-        self.preempt_count.fetch_sub(1, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn rcu_read_lock(&self) {
-        self.rcu_read_depth.fetch_add(1, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn rcu_read_unlock(&self) {
-        let mut current = self.rcu_read_depth.load(Ordering::SeqCst);
-        loop {
-            assert!(current > 0, "rcu_read_unlock underflow");
-            match self.rcu_read_depth.compare_exchange(
-                current,
-                current - 1,
-                Ordering::SeqCst,
-                Ordering::SeqCst,
-            ) {
-                Ok(_) => break,
-                Err(observed) => current = observed,
-            }
-        }
-    }
-
-    #[inline(always)]
-    pub unsafe fn set_preempt_count(&self, count: usize) {
-        self.preempt_count.store(count, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn contain_child(&self, pid: &RawPid) -> bool {
-        let children = self.children.read();
-        return children.contains(pid);
-    }
-
-    #[inline(always)]
-    pub fn flags(&self) -> &mut ProcessFlags {
-        return self.flags.get_mut();
-    }
-
-    #[inline(always)]
-    pub(crate) fn mark_visible_thread_accounted(&self) {
-        self.visible_thread_accounted.store(true, Ordering::Release);
-    }
-
-    #[inline(always)]
-    pub(crate) fn take_visible_thread_accounted(&self) -> bool {
-        self.visible_thread_accounted.swap(false, Ordering::AcqRel)
-    }
-
-    /// 请注意，这个值能在中断上下文中读取，但不能被中断上下文修改
-    /// 否则会导致死锁
-    #[inline(always)]
-    pub fn basic(&self) -> RwLockReadGuard<'_, ProcessBasicInfo> {
-        return self.basic.read_irqsave();
-    }
-
-    #[inline(always)]
-    pub fn set_name(&self, name: String) {
-        self.basic.write().set_name(name);
-    }
-
-    #[inline(always)]
-    pub fn set_pdeath_signal(&self, signal: Signal) {
-        self.pdeath_signal.store(signal, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn pdeath_signal(&self) -> Signal {
-        self.pdeath_signal.load(Ordering::SeqCst)
-    }
-
-    #[inline(always)]
-    pub fn no_new_privs(&self) -> usize {
-        if self.no_new_privs.load(Ordering::SeqCst) {
-            1
-        } else {
-            0
-        }
-    }
-
-    #[inline(always)]
-    pub fn set_no_new_privs(&self, value: bool) {
-        // Linux 语义：no_new_privs 一旦置位不可清除。
-        if value {
-            self.no_new_privs.store(true, Ordering::SeqCst);
-        }
-    }
-
-    #[inline(always)]
-    pub fn keepcaps(&self) -> bool {
-        self.keepcaps.load(Ordering::SeqCst)
-    }
-
-    #[inline(always)]
-    pub fn set_keepcaps(&self, value: bool) {
-        self.keepcaps.store(value, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn dumpable(&self) -> u8 {
-        self.dumpable.load(Ordering::SeqCst)
-    }
-
-    #[inline(always)]
-    pub fn set_dumpable(&self, value: u8) {
-        self.dumpable.store(value, Ordering::SeqCst)
-    }
-
-    #[inline(always)]
-    pub fn seccomp_mode(&self) -> seccomp::SeccompMode {
-        seccomp::SeccompMode::from(self.seccomp_mode.load(Ordering::Relaxed))
-    }
-
-    #[inline(always)]
-    pub fn set_seccomp_mode(&self, mode: seccomp::SeccompMode) {
-        self.seccomp_mode.store(mode as u8, Ordering::SeqCst);
-    }
-
-    #[inline(always)]
-    pub fn seccomp_filter_lock(&self) -> SpinLockGuard<'_, Option<Arc<seccomp::SeccompFilter>>> {
-        self.seccomp_filter.lock()
-    }
-
-    #[inline(always)]
-    pub fn basic_mut(&self) -> RwLockWriteGuard<'_, ProcessBasicInfo> {
-        return self.basic.write_irqsave();
-    }
-
-    /// # 获取arch info的锁，同时关闭中断
-    #[inline(always)]
-    pub fn arch_info_irqsave(&self) -> SpinLockGuard<'_, ArchPCBInfo> {
-        return self.arch_info.lock_irqsave();
-    }
-
-    /// # 获取arch info的锁，但是不关闭中断
-    ///
-    /// 由于arch info在进程切换的时候会使用到，
-    /// 因此在中断上下文外，获取arch info 而不irqsave是不安全的.
-    ///
-    /// 只能在以下情况下使用这个函数：
-    /// - 在中断上下文中（中断已经禁用），获取arch info的锁。
-    /// - 刚刚创建新的pcb
-    #[inline(always)]
-    pub unsafe fn arch_info(&self) -> SpinLockGuard<'_, ArchPCBInfo> {
-        return self.arch_info.lock();
-    }
-
-    #[inline(always)]
-    pub fn kernel_stack(&self) -> RwLockReadGuard<'_, KernelStack> {
-        return self.kernel_stack.read();
-    }
-
-    pub unsafe fn kernel_stack_force_ref(&self) -> &KernelStack {
-        self.kernel_stack.force_get_ref()
-    }
-
-    #[inline(always)]
-    #[allow(dead_code)]
-    pub fn kernel_stack_mut(&self) -> RwLockWriteGuard<'_, KernelStack> {
-        return self.kernel_stack.write();
-    }
-
-    #[inline(always)]
-    pub fn sched_info(&self) -> &ProcessSchedulerInfo {
-        return &self.sched_info;
-    }
-
-    pub fn sig_altstack(&self) -> RwLockReadGuard<'_, SigStackArch> {
-        self.sig_altstack.read_irqsave()
-    }
-
-    pub fn sig_altstack_mut(&self) -> RwLockWriteGuard<'_, SigStackArch> {
-        self.sig_altstack.write_irqsave()
-    }
-
-    #[inline(always)]
-    pub fn worker_private(&self) -> SpinLockGuard<'_, Option<WorkerPrivate>> {
-        return self.worker_private.lock();
-    }
-
-    #[inline(always)]
-    pub fn raw_pid(&self) -> RawPid {
-        return self.pid.load(Ordering::Acquire);
-    }
-
-    #[inline(always)]
-    pub fn raw_tgid(&self) -> RawPid {
-        return self.tgid;
-    }
-
-    #[inline(always)]
-    pub fn fs_struct(&self) -> Arc<FsStruct> {
-        self.fs.read().clone()
-    }
-
-    pub fn fs_struct_mut(&self) -> RwLockWriteGuard<'_, Arc<FsStruct>> {
-        self.fs.write()
-    }
-
-    #[inline(always)]
-    pub fn fs_struct_is_shared(&self) -> bool {
-        Arc::strong_count(&*self.fs.read()) > 1
-    }
-
-    pub fn set_fs_struct(&self, fs: Arc<FsStruct>) -> Arc<FsStruct> {
-        let mut guard = self.fs.write();
-        let old = guard.clone();
-        *guard = fs;
-        old
-    }
-
-    pub fn pwd_inode(&self) -> Arc<dyn IndexNode> {
-        self.fs.read().pwd()
-    }
-
-    /// 获取文件描述符表的Arc指针
-    #[inline(always)]
-    pub fn fd_table(&self) -> Arc<RwSem<FileDescriptorVec>> {
-        return self.basic.read().try_fd_table().unwrap();
-    }
-
-    #[inline(always)]
-    pub fn cred(&self) -> Arc<Cred> {
-        self.cred.load()
-    }
-
-    /// 原子替换当前进程的凭据集（cred）
-    ///
-    /// - 使用 irqsave 写锁保证并发安全
-    /// - 返回 Result 以便调用方在需要时扩展错误处理
-    pub fn set_cred(&self, new: Arc<Cred>) -> Result<(), SystemError> {
-        let _task_guard = self.task_lock.lock_irqsave();
-        self.cred.store_deferred(new);
-        Ok(())
-    }
-
-    pub fn set_execute_path(&self, path: String) {
-        *self.executable_path.write() = path;
-    }
-
-    pub fn execute_path(&self) -> String {
-        self.executable_path.read().clone()
-    }
-
-    /// 获取 /proc/<pid>/cmdline 的原始字节序列（argv 以 '\0' 分隔）。
-    #[inline(always)]
-    pub fn cmdline_bytes(&self) -> Vec<u8> {
-        self.cmdline.read().clone()
-    }
-
-    /// 直接设置 cmdline（用于 fork 继承等场景）。
-    #[inline(always)]
-    pub fn set_cmdline_bytes(&self, data: Vec<u8>) {
-        *self.cmdline.write() = data;
-    }
-
-    /// 在 exec 成功后写入 argv（Linux 语义：每个参数以 '\0' 结尾，整体通常也以 '\0' 结尾）。
-    #[inline(never)]
-    pub fn set_cmdline_from_argv(&self, argv: &[CString]) {
-        let mut buf: Vec<u8> = Vec::new();
-        for arg in argv {
-            buf.extend_from_slice(arg.as_bytes());
-            buf.push(0);
-        }
-        *self.cmdline.write() = buf;
-    }
-
-    pub fn real_parent_pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        return self.real_parent_pcb.read_irqsave().upgrade();
-    }
-
-    pub fn ptracer_pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        ptrace::ptracer_of(&self.self_ref.upgrade()?)
-    }
-
-    pub fn is_ptraced(&self) -> bool {
-        ptrace::is_ptraced(self)
-    }
-
-    pub fn ptraced_pids(&self) -> Vec<RawPid> {
-        let Some(this) = self.self_ref.upgrade() else {
-            return Vec::new();
-        };
-        ptrace::tracees_of(&this)
-    }
-
-    pub fn fork_parent_pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        self.fork_parent_pcb.read_irqsave().upgrade()
-    }
-
-    pub fn wait_parent_pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        self.wait_parent_pcb.read_irqsave().upgrade()
-    }
-
-    /// 判断当前进程是否是全局的init进程
-    pub fn is_global_init(&self) -> bool {
-        self.task_tgid_vnr().unwrap() == RawPid(1)
-    }
-
-    /// 根据文件描述符序号，获取socket相应的IndexNode的Arc指针
-    ///
-    /// this is a helper function
-    ///
-    /// ## 参数
-    ///
-    /// - `fd` 文件描述符序号
-    ///
-    /// ## 返回值
-    ///
-    /// 如果fd对应的文件是一个socket，那么返回这个socket相应的IndexNode的Arc指针，否则返回错误码
-    ///
-    /// # 注意
-    /// 因为底层的Socket中可能包含泛型，经过类型擦除转换成Arc<dyn Socket>的时候内部的泛型信息会丢失;
-    /// 因此这里返回Arc<dyn IndexNode>，可在外部直接通过 `as_socket()` 转换成 `Option<&dyn Socket>`;
-    /// 因为内部已经经过检查，因此在外部可以直接 `unwarp` 来获取 `&dyn Socket`
-    pub fn get_socket_inode(&self, fd: i32) -> Result<Arc<dyn IndexNode>, SystemError> {
-        let f = ProcessManager::current_pcb()
-            .fd_table()
-            .read()
-            .get_file_by_fd(fd)
-            .ok_or({
-                // log::warn!("get_socket: fd {} not found", fd);
-                SystemError::EBADF
-            })?;
-
-        if f.file_type() != FileType::Socket {
-            return Err(SystemError::EBADF);
-        }
-
-        let inode = f.inode();
-        // log::info!("get_socket: fd {} is a socket", fd);
-        if let Some(_sock) = inode.as_socket() {
-            // log::info!("{:?}", sock);
-            return Ok(inode);
-        }
-
-        Err(SystemError::EBADF)
-    }
-
-    fn is_alive_reparent_target(pcb: &Arc<ProcessControlBlock>) -> bool {
-        !pcb.flags().contains(ProcessFlags::EXITING)
-            && !pcb.is_exited()
-            && !pcb.is_zombie()
-            && !pcb.is_dead()
-    }
-
-    fn find_alive_thread_in_group(
-        pcb: Arc<ProcessControlBlock>,
-    ) -> Option<Arc<ProcessControlBlock>> {
-        ProcessManager::thread_group_tasks_snapshot(pcb)
-            .into_iter()
-            .find(ProcessControlBlock::is_alive_reparent_target)
-    }
-
-    fn find_alive_thread_reaper(
-        exiting: &Arc<ProcessControlBlock>,
-    ) -> Option<Arc<ProcessControlBlock>> {
-        ProcessManager::thread_group_tasks_snapshot(exiting.clone())
-            .into_iter()
-            .find(|task| {
-                !Arc::ptr_eq(task, exiting) && ProcessControlBlock::is_alive_reparent_target(task)
-            })
-    }
-
-    fn child_wait_parent_is(
-        child: &Arc<ProcessControlBlock>,
-        parent: &Arc<ProcessControlBlock>,
-    ) -> bool {
-        child
-            .wait_parent_pcb()
-            .as_ref()
-            .map(|wait_parent| Arc::ptr_eq(wait_parent, parent))
-            .unwrap_or(false)
-    }
-
-    fn link_child_to_parent_list(
-        child: &Arc<ProcessControlBlock>,
-        parent: &Arc<ProcessControlBlock>,
-    ) {
-        let child_vpid = child
-            .task_pid_nr_ns(PidType::PID, Some(parent.active_pid_ns()))
-            .unwrap_or(RawPid::new(0));
-        if child_vpid.data() == 0 {
-            return;
-        }
-
-        let mut children = parent.children.write_irqsave();
-        if !children.contains(&child_vpid) {
-            children.push(child_vpid);
-        }
-    }
-
-    pub(crate) fn reparent_child_to(
-        child: &Arc<ProcessControlBlock>,
-        new_parent: &Arc<ProcessControlBlock>,
-    ) {
-        let _relation_guard = PTRACE_RELATION_LOCK.lock_irqsave();
-        ProcessControlBlock::reparent_child_to_locked(child, new_parent);
-    }
-
-    fn reparent_child_to_locked(
-        child: &Arc<ProcessControlBlock>,
-        new_parent: &Arc<ProcessControlBlock>,
-    ) {
-        *child.parent_pcb.write_irqsave() = Arc::downgrade(new_parent);
-        *child.real_parent_pcb.write_irqsave() = Arc::downgrade(new_parent);
-        *child.wait_parent_pcb.write_irqsave() = Arc::downgrade(new_parent);
-        *child.fork_parent_pcb.write_irqsave() = Arc::downgrade(new_parent);
-
-        let parent_pid_in_child_ns = new_parent
-            .task_pid_nr_ns(PidType::PID, Some(child.active_pid_ns()))
-            .unwrap_or(RawPid::new(0));
-        child.basic.write_irqsave().ppid = parent_pid_in_child_ns;
-
-        ProcessControlBlock::link_child_to_parent_list(child, new_parent);
-
-        // Linux reparent_leader() notifies the new parent when a zombie child
-        // becomes wait-visible through reparenting. DragonOS keeps per-task wait
-        // queues, so wake both the concrete new parent and its group leader.
-        if child.is_zombie() {
-            ProcessManager::wake_wait_parent(new_parent);
-        }
-    }
-
-    pub(crate) fn reparent_child_links_from_thread_group(
-        child: &Arc<ProcessControlBlock>,
-        old_tgid: RawPid,
-        new_parent: &Arc<ProcessControlBlock>,
-    ) {
-        let should_reparent = child
-            .parent_pcb()
-            .as_ref()
-            .map(|parent| parent.tgid == old_tgid)
-            .unwrap_or(false)
-            || child
-                .real_parent_pcb()
-                .as_ref()
-                .map(|parent| parent.tgid == old_tgid)
-                .unwrap_or(false)
-            || child
-                .wait_parent_pcb()
-                .as_ref()
-                .map(|parent| parent.tgid == old_tgid)
-                .unwrap_or(false);
-
-        if should_reparent {
-            ProcessControlBlock::reparent_child_to(child, new_parent);
-        }
-    }
-
-    fn collect_children_for_exit(
-        exiting: &Arc<ProcessControlBlock>,
-    ) -> Vec<Arc<ProcessControlBlock>> {
-        // Caller holds PTRACE_RELATION_LOCK so collecting children, selecting a
-        // reaper, and moving children form one tasklist-lock-like transaction.
-        let mut result = Vec::new();
-        let mut seen = Vec::new();
-
-        let mut push_reparent_child =
-            |child: Arc<ProcessControlBlock>, result: &mut Vec<Arc<ProcessControlBlock>>| {
-                if seen.iter().any(|pid| *pid == child.raw_pid()) {
-                    return;
-                }
-                seen.push(child.raw_pid());
-                result.push(child);
-            };
-
-        for owner in ProcessManager::thread_group_tasks_snapshot(exiting.clone()) {
-            let owner_ns = owner.active_pid_ns();
-            let pids_to_rehome: Vec<RawPid> = if Arc::ptr_eq(&owner, exiting) {
-                let mut children = owner.children.write_irqsave();
-                core::mem::take(&mut *children)
-            } else {
-                let mut children = owner.children.write_irqsave();
-                let mut removed = Vec::new();
-                children.retain(|pid| {
-                    let should_remove = ProcessManager::find_task_by_pid_ns(*pid, &owner_ns)
-                        .as_ref()
-                        .map(|child| ProcessControlBlock::child_wait_parent_is(child, exiting))
-                        .unwrap_or(false);
-                    if should_remove {
-                        removed.push(*pid);
-                    }
-                    !should_remove
-                });
-                removed
-            };
-
-            for pid in pids_to_rehome {
-                let Some(child) = ProcessManager::find_task_by_pid_ns(pid, &owner_ns) else {
-                    continue;
-                };
-
-                if let Some(wait_parent) = child.wait_parent_pcb() {
-                    if !Arc::ptr_eq(&wait_parent, exiting)
-                        && wait_parent.tgid == exiting.tgid
-                        && ProcessControlBlock::is_alive_reparent_target(&wait_parent)
-                    {
-                        ProcessControlBlock::reparent_child_to_locked(&child, &wait_parent);
-                        continue;
-                    }
-                }
-
-                push_reparent_child(child, &mut result);
-            }
-        }
-
-        result
-    }
-
-    /// 当前进程退出时，让其他线程、subreaper 或 init 收养其子进程。
-    unsafe fn adopt_childen(&self) -> Result<(), SystemError> {
-        let exiting = self.self_ref.upgrade().ok_or(SystemError::ESRCH)?;
-        let _relation_guard = PTRACE_RELATION_LOCK.lock_irqsave();
-        let children = ProcessControlBlock::collect_children_for_exit(&exiting);
-
-        if children.is_empty() {
-            return Ok(());
-        }
-
-        self.notify_parent_exit_for_children(&children);
-
-        if let Some(reaper) = ProcessControlBlock::find_alive_thread_reaper(&exiting) {
-            for child in children {
-                ProcessControlBlock::reparent_child_to_locked(&child, &reaper);
-            }
-            return Ok(());
-        }
-
-        let exiting_ns = exiting.active_pid_ns();
-        let init_pcb = ProcessManager::find_task_by_pid_ns(RawPid(1), &exiting_ns)
-            .ok_or(SystemError::ECHILD)?;
-
-        // 如果当前进程是 namespace 的 init，则由父进程所在 pidns 的 init 去收养。
-        if Arc::ptr_eq(&exiting, &init_pcb) {
-            if let Some(parent_pcb) = self.real_parent_pcb() {
-                assert!(
-                    !Arc::ptr_eq(&parent_pcb, &init_pcb),
-                    "adopt_childen: parent_pcb is init_pcb, pid: {}",
-                    self.raw_pid()
-                );
-                let parent_init =
-                    ProcessManager::find_task_by_pid_ns(RawPid(1), &parent_pcb.active_pid_ns());
-                if let Some(parent_init) = parent_init {
-                    for child in children {
-                        ProcessControlBlock::reparent_child_to_locked(&child, &parent_init);
-                    }
-                }
-            }
-            return Ok(());
-        }
-
-        // 常规情况：优先 reparent 到“最近祖先 subreaper”，否则 reparent 到 init。
-        let mut reaper: Arc<ProcessControlBlock> = init_pcb.clone();
-        let mut cursor = self.real_parent_pcb();
-        while let Some(p) = cursor {
-            // Linux 语义：child_subreaper 是线程组级；这里统一按线程组 leader 判定。
-            let leader = {
-                let ti = p.threads_read_irqsave();
-                ti.group_leader().unwrap_or_else(|| p.clone())
-            };
-
-            let visible_in_exiting_ns = leader
-                .task_pid_nr_ns(PidType::PID, Some(exiting_ns.clone()))
-                .is_some_and(|pid| pid.data() != 0);
-            // Linux uses task_pid(reaper)->level == ns_level. DragonOS exposes
-            // this through active pidns level plus visibility in exiting_ns.
-            if !visible_in_exiting_ns || leader.active_pid_ns().level() != exiting_ns.level() {
-                break;
-            }
-
-            if leader.sig_info_irqsave().is_child_subreaper() {
-                if let Some(alive) = ProcessControlBlock::find_alive_thread_in_group(leader.clone())
-                {
-                    reaper = alive;
-                    break;
-                }
-            }
-
-            if Arc::ptr_eq(&leader, &init_pcb) {
-                break;
-            }
-
-            cursor = leader.real_parent_pcb();
-        }
-
-        for child in children {
-            ProcessControlBlock::reparent_child_to_locked(&child, &reaper);
-        }
-
-        Ok(())
-    }
-
-    fn notify_parent_exit_for_children(&self, children: &[Arc<ProcessControlBlock>]) {
-        for child in children {
-            let sig = child.pdeath_signal();
-            if sig == Signal::INVALID {
-                continue;
-            }
-            if let Err(e) = crate::ipc::kill::send_signal_to_pcb(child.clone(), sig) {
-                warn!(
-                    "adopt_childen: failed to deliver pdeath_signal {:?} to child {:?}: {:?}",
-                    sig,
-                    child.raw_pid(),
-                    e
-                );
-            }
-        }
-    }
-
-    /// 生成进程的名字
-    pub fn generate_name(program_path: &str) -> String {
-        // Extract just the basename from the program path
-        let name = program_path.split('/').next_back().unwrap_or(program_path);
-        name.to_string()
-    }
-
-    pub fn sig_info_irqsave(&self) -> RwLockReadGuard<'_, ProcessSignalInfo> {
-        self.sig_info.read_irqsave()
-    }
-
-    pub fn sig_info_upgradable(&self) -> RwLockUpgradableGuard<'_, ProcessSignalInfo> {
-        self.sig_info.upgradeable_read_irqsave()
-    }
-
-    pub fn try_siginfo_irqsave(&self, times: u8) -> Option<RwLockReadGuard<'_, ProcessSignalInfo>> {
-        for _ in 0..times {
-            if let Some(r) = self.sig_info.try_read_irqsave() {
-                return Some(r);
-            }
-        }
-
-        return None;
-    }
-
-    pub fn sig_info_mut(&self) -> RwLockWriteGuard<'_, ProcessSignalInfo> {
-        self.sig_info.write_irqsave()
-    }
-
-    pub fn is_active_vfork(&self) -> bool {
-        self.thread.read_irqsave().vfork_done.is_some()
-    }
-
-    /// 获取 rseq 状态的只读引用
-    #[inline]
-    pub fn rseq_state(&self) -> RwLockReadGuard<'_, rseq::RseqState> {
-        self.rseq_state.read_irqsave()
-    }
-
-    /// 获取 rseq 状态的可变引用
-    #[inline]
-    pub fn rseq_state_mut(&self) -> RwLockWriteGuard<'_, rseq::RseqState> {
-        self.rseq_state.write_irqsave()
-    }
-
-    pub fn try_siginfo_mut(&self, times: u8) -> Option<RwLockWriteGuard<'_, ProcessSignalInfo>> {
-        for _ in 0..times {
-            if let Some(r) = self.sig_info.try_write_irqsave() {
-                return Some(r);
-            }
-        }
-
-        return None;
-    }
-
-    /// 判断当前进程是否有未处理的信号
-    pub fn has_pending_signal(&self) -> bool {
-        let sig_info = self.sig_info_irqsave();
-        let has_pending_thread = sig_info.sig_pending().has_pending();
-        drop(sig_info);
-        if has_pending_thread {
-            return true;
-        }
-        // also check shared-pending in sighand
-        let shared = self.sighand().shared_pending_signal();
-        return !shared.is_empty();
-    }
-
-    /// 根据 pcb 的 flags 判断当前进程是否有未处理的信号
-    pub fn has_pending_signal_fast(&self) -> bool {
-        self.flags.get().contains(ProcessFlags::HAS_PENDING_SIGNAL)
-    }
-
-    /// 检查当前进程是否有未被阻塞的待处理信号。
-    ///
-    /// 注：该函数较慢，因此需要与 has_pending_signal_fast 一起使用。
-    pub fn has_pending_not_masked_signal(&self) -> bool {
-        let sig_info = self.sig_info_irqsave();
-        let blocked: SigSet = *sig_info.sig_blocked();
-        let mut pending: SigSet = sig_info.sig_pending().signal();
-        drop(sig_info);
-        // 同时检查 shared_pending
-        pending |= self.sighand().shared_pending_signal();
-        pending.remove(blocked);
-        // log::debug!(
-        //     "pending and not masked:{:?}, masked: {:?}",
-        //     pending,
-        //     blocked
-        // );
-        let has_not_masked = !pending.is_empty();
-        return has_not_masked;
-    }
-
-    #[inline(always)]
-    pub fn get_robust_list(&self) -> RwLockReadGuard<'_, Option<RobustListHead>> {
-        return self.robust_list.read_irqsave();
-    }
-
-    #[inline(always)]
-    pub fn set_robust_list(&self, new_robust_list: Option<RobustListHead>) {
-        *self.robust_list.write_irqsave() = new_robust_list;
-    }
-
-    #[inline(always)]
-    pub fn alarm_timer_irqsave(&self) -> SpinLockGuard<'_, Option<AlarmTimer>> {
-        return self.alarm_timer.lock_irqsave();
-    }
-
-    #[inline(always)]
-    pub fn itimers_irqsave(&self) -> SpinLockGuard<'_, ProcessItimers> {
-        return self.itimers.lock_irqsave();
-    }
-
-    pub fn posix_timers_irqsave(&self) -> SpinLockGuard<'_, posix_timer::ProcessPosixTimers> {
-        return self.posix_timers.lock_irqsave();
-    }
-
-    /// 清理当前进程/线程的定时器
-    ///
-    /// 参考 Linux do_exit() 中的定时器清理逻辑：
-    /// ```c
-    /// if (group_dead) {
-    ///     hrtimer_cancel(&tsk->signal->real_timer);
-    ///     exit_itimers(tsk);
-    /// }
-    /// ```
-    ///
-    /// DragonOS 中 alarm_timer 是 per-PCB 的，因此每个线程退出时
-    /// 都需要取消自己的 alarm timer。itimers 和 posix_timers 仅在
-    /// 线程组 leader 退出时（group_dead）清理。
-    fn exit_timers(&self) {
-        // 1. 取消当前线程的 alarm timer
-        if let Some(alarm) = self.alarm_timer.lock_irqsave().take() {
-            alarm.cancel();
-        }
-
-        let group_dead = self.is_thread_group_leader();
-        if group_dead {
-            // 2. 取消 ITIMER_REAL
-            if let Some(real_itimer) = self.itimers.lock_irqsave().real.take() {
-                real_itimer.timer.cancel();
-            }
-
-            // 3. 删除所有 POSIX interval timers
-            let mut posix_timers = self.posix_timers.lock_irqsave();
-            let timer_ids: alloc::vec::Vec<i32> = posix_timers.timer_ids().collect();
-            let self_arc = self.self_ref.upgrade();
-            if let Some(ref pcb) = self_arc {
-                for id in timer_ids {
-                    let _ = posix_timers.delete(pcb, id);
-                }
-            }
-        }
-    }
-
-    /// Exit fd table when process exit
-    fn exit_files(&self) {
-        // 关闭文件描述符表
-        // 这里这样写的原因是避免某些inode在关闭时需要访问当前进程的basic，导致死锁
-        let mut guard = self.basic.write_irqsave();
-        let old = guard.set_fd_table(None);
-        drop(guard);
-        drop(old)
-    }
-
-    pub fn children_read_irqsave(&self) -> RwLockReadGuard<'_, Vec<RawPid>> {
-        self.children.read_irqsave()
-    }
-
-    pub fn threads_read_irqsave(&self) -> RwLockReadGuard<'_, ThreadInfo> {
-        self.thread.read_irqsave()
-    }
-
-    pub fn threads_write_irqsave(&self) -> RwLockWriteGuard<'_, ThreadInfo> {
-        self.thread.write_irqsave()
-    }
-
-    pub fn restart_block(&self) -> SpinLockGuard<'_, Option<RestartBlock>> {
-        self.restart_block.lock()
-    }
-
-    pub fn set_restart_fn(
-        &self,
-        restart_block: Option<RestartBlock>,
-    ) -> Result<usize, SystemError> {
-        *self.restart_block.lock() = restart_block;
-        return Err(SystemError::ERESTART_RESTARTBLOCK);
-    }
-
-    pub fn parent_pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        self.parent_pcb.read().upgrade()
-    }
-
-    pub fn is_exited(&self) -> bool {
-        self.sched_info.state().is_exited()
-    }
-
-    /// DragonOS sets `ProcessState::Exited` after the task has left user-visible
-    /// execution, before `ExitState::Zombie`; it is no longer a live thread-group
-    /// member for wait/pidfd readiness.
-    pub(crate) fn is_live_thread_group_member(&self) -> bool {
-        !self.is_exited() && !self.is_zombie() && !self.is_dead()
-    }
-
-    pub fn exit_code(&self) -> Option<usize> {
-        self.sched_info.state().exit_code()
-    }
-
-    pub fn exit_state(&self) -> ExitState {
-        ExitState::from_u8(self.exit_state.load(Ordering::Acquire))
-    }
-
-    pub fn is_zombie(&self) -> bool {
-        self.exit_state() == ExitState::Zombie
-    }
-
-    pub fn is_dead(&self) -> bool {
-        self.exit_state() == ExitState::Dead
-    }
-
-    pub fn set_exit_state_zombie(&self) {
-        self.exit_state
-            .store(ExitState::Zombie as u8, Ordering::Release);
-    }
-
-    pub fn try_mark_dead_from_zombie(&self) -> bool {
-        self.exit_state
-            .compare_exchange(
-                ExitState::Zombie as u8,
-                ExitState::Dead as u8,
-                Ordering::AcqRel,
-                Ordering::Acquire,
-            )
-            .is_ok()
-    }
-
-    pub fn mark_exiting(&self) {
-        self.flags().insert(ProcessFlags::EXITING);
-        fence(Ordering::SeqCst);
-    }
-
-    /// 获取进程的namespace代理
-    pub fn nsproxy(&self) -> Arc<NsProxy> {
-        self.nsproxy.load()
-    }
-
-    /// 设置进程的namespace代理
-    ///
-    /// ## 参数
-    /// - `nsproxy` : 新的namespace代理
-    pub fn set_nsproxy(&self, nsproxy: Arc<NsProxy>) {
-        let _task_guard = self.task_lock.lock_irqsave();
-        self.nsproxy.store_deferred(nsproxy);
-    }
-
-    pub fn task_cgroup_ref(&self) -> TaskCgroupRef {
-        self.task_cgroup.read().clone()
-    }
-
-    pub fn task_cgroup_node(&self) -> Arc<CgroupNode> {
-        self.task_cgroup.read().node()
-    }
-
-    /// 设置任务所属的 cgroup 节点
-    ///
-    /// # 安全性
-    ///
-    /// 调用者必须持有 `cgroup_accounting_lock` 以避免死锁和竞态条件
-    pub fn set_task_cgroup_node(&self, node: Arc<CgroupNode>) {
-        // 先使用读锁获取 old 节点
-        let old = {
-            let task_cgroup = self.task_cgroup.read();
-            let old = task_cgroup.node();
-            if Arc::ptr_eq(&old, &node) {
-                return;
-            }
-            old
-        }; // 释放读锁
-
-        // 在不持有 task_cgroup 锁的情况下执行迁移；调用者必须持有
-        // cgroup_accounting_lock，保证可见成员和 pids 计费一起切换。
-        let pid = self.raw_pid();
-        old.remove_task(pid);
-        node.add_task(pid);
-        CgroupNode::transfer_pids_charge(&old, &node, 1);
-
-        // 使用写锁更新 task_cgroup
-        let mut task_cgroup = self.task_cgroup.write();
-        *task_cgroup = TaskCgroupRef::new(node);
-    }
-
-    /// 仅用于 fork 时设置任务的 cgroup 节点
-    ///
-    /// # 安全性
-    ///
-    /// 调用者必须持有 `cgroup_accounting_lock`
-    ///
-    /// # 注意
-    ///
-    /// 此函数只更新 task_cgroup 引用，不会调用 add_task()。
-    /// add_task() 会在后续的 ProcessManager::add_pcb() 中被调用。
-    pub fn set_task_cgroup_node_for_fork(&self, node: Arc<CgroupNode>) {
-        *self.task_cgroup.write() = TaskCgroupRef::new(node);
-    }
-
-    pub fn is_thread_group_leader(&self) -> bool {
-        self.exit_signal.load(Ordering::SeqCst) >= 0
-    }
-
-    pub(crate) fn thread_group_has_live_nonleader_threads(&self) -> bool {
-        if !self.is_thread_group_leader() {
-            return false;
-        }
-
-        self.threads_read_irqsave()
-            .group_tasks_clone()
-            .into_iter()
-            .filter_map(|task| task.upgrade())
-            .any(|task| task.is_live_thread_group_member())
-    }
-
-    /// 唤醒等待在本进程 `wait_queue` 上的所有等待者
-    pub fn wake_all_waiters(&self) {
-        self.wait_queue
-            .wakeup_all(Some(ProcessState::Blocked(true)))
-    }
-}
-
-impl Drop for ProcessControlBlock {
-    fn drop(&mut self) {
-        let irq_guard = unsafe { CurrentIrqArch::save_and_disable_irq() };
-        // log::debug!("Drop ProcessControlBlock: pid: {}", self.raw_pid(),);
-        self.__exit_signal();
-        self.sighand().detach_task_ref();
-        // 新的 ProcFS 是动态的，进程目录会在访问时按需创建
-        // 不再需要显式注册/注销进程
-        if let Some(ppcb) = self.parent_pcb.read_irqsave().upgrade() {
-            ppcb.children
-                .write_irqsave()
-                .retain(|pid| *pid != self.raw_pid());
-        }
-
-        // log::debug!("Drop pid: {:?}", self.pid());
-        drop(irq_guard);
-    }
-}
-
-/// 线程信息
-#[derive(Debug)]
-pub struct ThreadInfo {
-    // 来自用户空间记录用户线程id的地址，在该线程结束时将该地址置0以通知父进程
-    clear_child_tid: Option<VirtAddr>,
-    set_child_tid: Option<VirtAddr>,
-
-    vfork_done: Option<Arc<Completion>>,
-    /// 线程组的组长
-    group_leader: Weak<ProcessControlBlock>,
-
-    /// 当前线程为组长时，该字段存储组内所有线程的pcb
-    group_tasks: Vec<Weak<ProcessControlBlock>>,
-}
-
-impl Default for ThreadInfo {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl ThreadInfo {
-    pub fn new() -> Self {
-        Self {
-            clear_child_tid: None,
-            set_child_tid: None,
-            vfork_done: None,
-            group_leader: Weak::default(),
-            group_tasks: Vec::new(),
-        }
-    }
-
-    pub fn group_leader(&self) -> Option<Arc<ProcessControlBlock>> {
-        return self.group_leader.upgrade();
-    }
-
-    /// 返回线程组中除组长外的线程列表（由组长维护）。
-    ///
-    /// 说明：该列表存储为 `Weak`，调用者应当在使用前 `upgrade()` 并处理失败场景。
-    pub fn group_tasks_clone(&self) -> Vec<Weak<ProcessControlBlock>> {
-        self.group_tasks.clone()
-    }
-
-    pub fn thread_group_empty(&self) -> bool {
-        let group_leader = self.group_leader();
-        if let Some(leader) = group_leader {
-            if Arc::ptr_eq(&leader, &ProcessManager::current_pcb()) {
-                if self.group_tasks.is_empty() {
-                    return true;
-                }
-                // 仅当组内存在“存活线程”时才返回 false
-                for weak in &self.group_tasks {
-                    if let Some(task) = weak.upgrade() {
-                        if Arc::ptr_eq(&task, &ProcessManager::current_pcb()) {
-                            continue;
-                        }
-                        if !task.is_exited() && !task.is_dead() && !task.is_zombie() {
-                            return false;
-                        }
-                    }
-                }
-                return true;
-            }
-            return false;
-        }
-        return true;
-    }
-}
-
-/// 进程的基本信息
-///
-/// 这个结构体保存进程的基本信息，主要是那些不会随着进程的运行而经常改变的信息。
-#[derive(Debug)]
-pub struct ProcessBasicInfo {
-    /// 当前进程的父进程的pid
-    ppid: RawPid,
-    /// 进程的名字
-    name: String,
-
-    /// 当前进程的工作目录
-    cwd: String,
-
-    /// 用户地址空间
-    user_vm: Option<Arc<AddressSpace>>,
-
-    /// 文件描述符表
-    fd_table: Option<Arc<RwSem<FileDescriptorVec>>>,
-}
-
-impl ProcessBasicInfo {
-    #[inline(never)]
-    pub fn new(
-        ppid: RawPid,
-        name: String,
-        cwd: String,
-        user_vm: Option<Arc<AddressSpace>>,
-    ) -> RwLock<Self> {
-        let fd_table = Arc::new(RwSem::new(FileDescriptorVec::new()));
-        return RwLock::new(Self {
-            ppid,
-            name,
-            cwd,
-            user_vm,
-            fd_table: Some(fd_table),
-        });
-    }
-
-    pub fn ppid(&self) -> RawPid {
-        return self.ppid;
-    }
-
-    pub fn name(&self) -> &str {
-        return &self.name;
-    }
-
-    pub fn set_name(&mut self, name: String) {
-        self.name = name;
-    }
-
-    pub fn cwd(&self) -> String {
-        return self.cwd.clone();
-    }
-    pub fn set_cwd(&mut self, path: String) {
-        return self.cwd = path;
-    }
-
-    pub fn user_vm(&self) -> Option<Arc<AddressSpace>> {
-        return self.user_vm.clone();
-    }
-
-    pub unsafe fn set_user_vm(&mut self, user_vm: Option<Arc<AddressSpace>>) {
-        self.user_vm = user_vm;
-    }
-
-    /// Replace the task's user address space and return the old one without
-    /// dropping it while this lock is still held. The caller must preserve the
-    /// active_cpus and TLB-state ordering required by the mm switch/exit path.
-    pub unsafe fn replace_user_vm(
-        &mut self,
-        user_vm: Option<Arc<AddressSpace>>,
-    ) -> Option<Arc<AddressSpace>> {
-        let old = self.user_vm.take();
-        self.user_vm = user_vm;
-        old
-    }
-
-    pub fn try_fd_table(&self) -> Option<Arc<RwSem<FileDescriptorVec>>> {
-        return self.fd_table.clone();
-    }
-
-    #[inline]
-    pub fn fd_table_is_shared(&self) -> bool {
-        self.fd_table
-            .as_ref()
-            .map(|t| Arc::strong_count(t) > 1)
-            .unwrap_or(false)
-    }
-
-    pub fn set_fd_table(
-        &mut self,
-        fd_table: Option<Arc<RwSem<FileDescriptorVec>>>,
-    ) -> Option<Arc<RwSem<FileDescriptorVec>>> {
-        let old = self.fd_table.take();
-        self.fd_table = fd_table;
-        return old;
-    }
-}
-
-#[derive(Debug)]
-pub struct ProcessSchedulerInfo {
-    /// 当前进程所在的cpu
-    on_cpu: AtomicProcessorId,
-    /// 如果当前进程等待被迁移到另一个cpu核心上（也就是flags中的PF_NEED_MIGRATE被置位），
-    /// 该字段存储要被迁移到的目标处理器核心号
-    migrate_to: AtomicProcessorId,
-    state_atomic: AtomicU32,
-    pi_lock: SpinLock<PiProtected>,
-    /// 进程的调度优先级
-    // priority: SchedPriority,
-    /// 当前进程的虚拟运行时间
-    // virtual_runtime: AtomicIsize,
-    /// 由实时调度器管理的时间片
-    // rt_time_slice: AtomicIsize,
-    pub sched_stat: RwLock<SchedInfo>,
-    /// 调度策略（由 rq_lock / pi_lock 保护）
-    sched_policy: AtomicU8,
-    /// cfs调度实体
-    pub sched_entity: Arc<FairSchedEntity>,
-    pub on_rq: SpinLock<OnRq>,
-    placement: SpinLock<NewTaskPlacement>,
-
-    /// 由 rq_lock 保护
-    prio: AtomicI32,
-    static_prio: AtomicI32,
-    normal_prio: AtomicI32,
-}
-
-#[derive(Debug, Clone, Copy, Default)]
-pub(crate) struct NewTaskPlacement {
-    is_new_task: bool,
-    target_cpu_hint: Option<ProcessorId>,
-}
-
-#[derive(Debug, Default)]
-#[allow(dead_code)]
-pub struct SchedInfo {
-    /// 记录任务在特定 CPU 上运行的次数
-    pub pcount: usize,
-    /// 记录任务等待在运行队列上的时间
-    pub run_delay: usize,
-    /// 记录任务上次在 CPU 上运行的时间戳
-    pub last_arrival: u64,
-    /// 记录任务上次被加入到运行队列中的时间戳
-    pub last_queued: u64,
-}
-
-/// 由 pi_lock 保护的字段集合。
-#[derive(Debug)]
-pub struct PiProtected {
-    pub cpus_allowed: CpuMask,
-    pub nr_cpus_allowed: usize,
-}
-
-impl PiProtected {
-    pub fn new(cpus_allowed: CpuMask) -> Self {
-        let nr_cpus_allowed = cpus_allowed.iter_cpu().count();
-        Self {
-            cpus_allowed,
-            nr_cpus_allowed,
-        }
-    }
-
-    pub fn set_cpus_allowed(&mut self, new_mask: CpuMask) {
-        self.cpus_allowed = new_mask;
-        self.nr_cpus_allowed = self.cpus_allowed.iter_cpu().count();
-    }
-}
-
-impl ProcessSchedulerInfo {
-    fn default_cpus_allowed() -> CpuMask {
-        if crate::smp::cpu::smp_cpu_manager_initialized() {
-            return crate::smp::cpu::smp_cpu_manager().possible_cpus().clone();
-        }
-
-        // 进程管理初始化早于 SMP 拓扑初始化。此时先退化为“当前引导 CPU 可运行”，
-        // 待后续显式初始化（如 idle/per-cpu kthread）再收敛到更精确的 mask。
-        CpuMask::from_cpu(current_cpu_id())
-    }
-
-    #[inline(never)]
-    pub fn new(on_cpu: Option<ProcessorId>) -> Self {
-        let cpu_id = on_cpu.unwrap_or(ProcessorId::INVALID);
-        let cpus_allowed = Self::default_cpus_allowed();
-        return Self {
-            on_cpu: AtomicProcessorId::new(cpu_id),
-            migrate_to: AtomicProcessorId::new(ProcessorId::INVALID),
-            state_atomic: AtomicU32::new(ProcessState::Blocked(false).to_u32()),
-            pi_lock: SpinLock::new(PiProtected::new(cpus_allowed)),
-            // virtual_runtime: AtomicIsize::new(0),
-            // rt_time_slice: AtomicIsize::new(0),
-            // priority: SchedPriority::new(100).unwrap(),
-            sched_stat: RwLock::new(SchedInfo::default()),
-            sched_policy: AtomicU8::new(SchedPolicy::CFS.to_u8()),
-            sched_entity: FairSchedEntity::new(),
-            on_rq: SpinLock::new(OnRq::None),
-            placement: SpinLock::new(NewTaskPlacement::default()),
-            prio: AtomicI32::new(MAX_PRIO - 20),
-            static_prio: AtomicI32::new(MAX_PRIO - 20),
-            normal_prio: AtomicI32::new(MAX_PRIO - 20),
-        };
-    }
-
-    pub fn sched_entity(&self) -> Arc<FairSchedEntity> {
-        return self.sched_entity.clone();
-    }
-
-    pub fn on_cpu(&self) -> Option<ProcessorId> {
-        let on_cpu = self.on_cpu.load(Ordering::SeqCst);
-        if on_cpu == ProcessorId::INVALID {
-            return None;
-        } else {
-            return Some(on_cpu);
-        }
-    }
-
-    pub fn set_on_cpu(&self, on_cpu: Option<ProcessorId>) {
-        if let Some(cpu_id) = on_cpu {
-            self.on_cpu.store(cpu_id, Ordering::SeqCst);
-        } else {
-            self.on_cpu.store(ProcessorId::INVALID, Ordering::SeqCst);
-        }
-    }
-
-    #[allow(dead_code)]
-    pub(crate) fn placement_lock(&self) -> SpinLockGuard<'_, NewTaskPlacement> {
-        self.placement.lock_irqsave()
-    }
-
-    pub fn mark_new_task(&self, target_cpu_hint: Option<ProcessorId>) {
-        let mut guard = self.placement.lock_irqsave();
-        guard.is_new_task = true;
-        guard.target_cpu_hint = target_cpu_hint;
-    }
-
-    /// # 参数
-    /// - `allowed`: 调用者必须在 pi_lock 保护下预先读取 cpus_allowed 并传入，以避免重入 pi_lock
-    pub fn consume_new_task_target_cpu(
-        &self,
-        current_cpu: ProcessorId,
-        allowed: CpuMask,
-        default_selector: impl FnOnce(&CpuMask) -> Option<ProcessorId>,
-    ) -> Result<ProcessorId, SystemError> {
-        let mut placement = self.placement.lock_irqsave();
-        if !placement.is_new_task {
-            return Err(SystemError::EINVAL);
-        }
-        let selected_cpu = default_selector(&allowed).filter(|&cpu| {
-            allowed.get(cpu).unwrap_or(false)
-                && (!crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(cpu))
-        });
-        let target_cpu = if let Some(target_cpu) = placement.target_cpu_hint {
-            if allowed.get(target_cpu).unwrap_or(false)
-                && (!crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(target_cpu))
-            {
-                target_cpu
-            } else if let Some(selected_cpu) = selected_cpu {
-                selected_cpu
-            } else if allowed.get(current_cpu).unwrap_or(false)
-                && (!crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(current_cpu))
-            {
-                current_cpu
-            } else {
-                allowed
-                    .iter_cpu()
-                    .find(|&cpu| {
-                        !crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(cpu)
-                    })
-                    .ok_or(SystemError::EINVAL)?
-            }
-        } else if let Some(selected_cpu) = selected_cpu {
-            selected_cpu
-        } else if allowed.get(current_cpu).unwrap_or(false)
-            && (!crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(current_cpu))
-        {
-            current_cpu
-        } else {
-            allowed
-                .iter_cpu()
-                .find(|&cpu| !crate::smp::cpu::smp_cpu_manager_initialized() || cpu_is_online(cpu))
-                .ok_or(SystemError::EINVAL)?
-        };
-
-        placement.is_new_task = false;
-        placement.target_cpu_hint = None;
-        Ok(target_cpu)
-    }
-
-    pub fn is_new_task(&self) -> bool {
-        self.placement.lock_irqsave().is_new_task
-    }
-
-    pub fn migrate_to(&self) -> Option<ProcessorId> {
-        let migrate_to = self.migrate_to.load(Ordering::SeqCst);
-        if migrate_to == ProcessorId::INVALID {
-            None
-        } else {
-            Some(migrate_to)
-        }
-    }
-
-    pub fn set_migrate_to(&self, migrate_to: Option<ProcessorId>) {
-        self.migrate_to
-            .store(migrate_to.unwrap_or(ProcessorId::INVALID), Ordering::SeqCst);
-    }
-
-    pub fn state(&self) -> ProcessState {
-        ProcessState::from_u32(self.state_atomic.load(Ordering::Acquire))
-    }
-
-    pub fn set_state(&self, state: ProcessState) {
-        self.state_atomic.store(state.to_u32(), Ordering::Release);
-    }
-
-    /// 获取 pi_lock（SpinLock），同时关闭中断并禁用抢占。
-    /// 锁序：pi_lock 可嵌套 rq_lock（pi_lock → rq_lock），禁止反向。
-    pub fn pi_lock_irqsave(&self) -> SpinLockGuard<'_, PiProtected> {
-        self.pi_lock.lock_irqsave()
-    }
-
-    // pub fn virtual_runtime(&self) -> isize {
-    //     return self.virtual_runtime.load(Ordering::SeqCst);
-    // }
-
-    // pub fn set_virtual_runtime(&self, virtual_runtime: isize) {
-    //     self.virtual_runtime
-    //         .store(virtual_runtime, Ordering::SeqCst);
-    // }
-    // pub fn increase_virtual_runtime(&self, delta: isize) {
-    //     self.virtual_runtime.fetch_add(delta, Ordering::SeqCst);
-    // }
-
-    // pub fn rt_time_slice(&self) -> isize {
-    //     return self.rt_time_slice.load(Ordering::SeqCst);
-    // }
-
-    // pub fn set_rt_time_slice(&self, rt_time_slice: isize) {
-    //     self.rt_time_slice.store(rt_time_slice, Ordering::SeqCst);
-    // }
-
-    // pub fn increase_rt_time_slice(&self, delta: isize) {
-    //     self.rt_time_slice.fetch_add(delta, Ordering::SeqCst);
-    // }
-
-    /// 读取调度策略
-    #[inline]
-    pub fn policy(&self) -> SchedPolicy {
-        SchedPolicy::from_u8(self.sched_policy.load(Ordering::Relaxed))
-    }
-
-    /// 设置调度策略
-    #[inline]
-    pub fn set_policy(&self, policy: SchedPolicy) {
-        self.sched_policy.store(policy.to_u8(), Ordering::Relaxed);
-    }
-
-    /// 读取动态优先级
-    #[inline]
-    pub fn prio(&self) -> i32 {
-        self.prio.load(Ordering::Relaxed)
-    }
-
-    /// 设置动态优先级
-    #[inline]
-    pub fn set_prio(&self, val: i32) {
-        self.prio.store(val, Ordering::Relaxed);
-    }
-
-    /// 读取静态优先级
-    #[inline]
-    pub fn static_prio(&self) -> i32 {
-        self.static_prio.load(Ordering::Relaxed)
-    }
-
-    /// 设置静态优先级
-    #[inline]
-    pub fn set_static_prio(&self, val: i32) {
-        self.static_prio.store(val, Ordering::Relaxed);
-    }
-
-    /// 读取普通优先级
-    #[inline]
-    pub fn normal_prio(&self) -> i32 {
-        self.normal_prio.load(Ordering::Relaxed)
-    }
-
-    /// 设置普通优先级
-    #[inline]
-    pub fn set_normal_prio(&self, val: i32) {
-        self.normal_prio.store(val, Ordering::Relaxed);
-    }
-
-    /// 同时设置所有优先级
-    #[inline]
-    pub fn set_prio_all(&self, prio: i32, static_prio: i32, normal_prio: i32) {
-        self.prio.store(prio, Ordering::Relaxed);
-        self.static_prio.store(static_prio, Ordering::Relaxed);
-        self.normal_prio.store(normal_prio, Ordering::Relaxed);
-    }
-
-    pub fn cpus_allowed(&self) -> CpuMask {
-        self.pi_lock.lock_irqsave().cpus_allowed.clone()
-    }
-
-    pub fn nr_cpus_allowed(&self) -> usize {
-        self.pi_lock.lock_irqsave().nr_cpus_allowed
-    }
-
-    pub fn set_cpus_allowed(&self, cpus_allowed: CpuMask) {
-        self.pi_lock.lock_irqsave().set_cpus_allowed(cpus_allowed);
-    }
-}
-
-impl ProcessControlBlock {
-    #[inline]
-    pub(crate) fn debug_assert_fork_cpu_binding(&self) {
-        if cfg!(debug_assertions) {
-            let Some(on_cpu) = self.sched_info().on_cpu() else {
-                return;
-            };
-
-            debug_assert_eq!(
-                self.sched_info().sched_entity().cfs_rq().rq().cpu(),
-                on_cpu,
-                "fork target cpu and SE bound rq must stay consistent"
-            );
-        }
-    }
-}
-
-#[derive(Debug)]
-pub struct KernelStack {
-    stack: Option<AlignedBox<[u8; KernelStack::SIZE], { KernelStack::ALIGN }>>,
-    /// 标记该内核栈是否可以被释放
-    ty: KernelStackType,
-}
-
-#[derive(Debug)]
-pub enum KernelStackType {
-    KernelSpace(VirtAddr, PhysAddr),
-    Static,
-    Dynamic,
-}
-
-// 为什么需要这个锁?
-// alloc_from_kernel_space 使用该函数分配内核栈时，如果该函数被中断打断，
-// 而切换的任务使用dealloc_from_kernel_space回收内核栈，对
-// KernelMapper的可变引用获取将会失败造成错误
-static KSTACK_LOCK: SpinLock<()> = SpinLock::new(());
-
-unsafe fn alloc_from_kernel_space() -> (VirtAddr, PhysAddr) {
-    use crate::arch::MMArch;
-    use crate::mm::allocator::page_frame::{allocate_page_frames, PageFrameCount};
-    use crate::mm::kernel_mapper::KernelMapper;
-    use crate::mm::page::EntryFlags;
-    use crate::mm::MemoryManagementArch;
-
-    // Layout
-    // ---------------
-    // | KernelStack |
-    // | guard page  | size == KernelStack::SIZE
-    // | KernelStack |
-    // | guard page  |
-    // | ..........  |
-    // ---------------
-
-    let _guard = KSTACK_LOCK.try_lock_irqsave().unwrap();
-    let need_size = KernelStack::SIZE * 2;
-    let page_num = PageFrameCount::new(need_size.div_ceil(MMArch::PAGE_SIZE).next_power_of_two());
-
-    let (paddr, _count) = allocate_page_frames(page_num).expect("kernel stack alloc failed");
-
-    let guard_vaddr = MMArch::phys_2_virt(paddr).unwrap();
-    let _kstack_paddr = paddr + KernelStack::SIZE;
-    let kstack_vaddr = guard_vaddr + KernelStack::SIZE;
-
-    core::ptr::write_bytes(kstack_vaddr.data() as *mut u8, 0, KernelStack::SIZE);
-
-    let guard_flags = EntryFlags::new();
-
-    let mut kernel_mapper = KernelMapper::lock();
-    let kernel_mapper = kernel_mapper.as_mut().unwrap();
-
-    for i in 0..KernelStack::SIZE / MMArch::PAGE_SIZE {
-        let guard_page_vaddr = guard_vaddr + i * MMArch::PAGE_SIZE;
-        // Map the guard page
-        let flusher = kernel_mapper.remap(guard_page_vaddr, guard_flags).unwrap();
-        flusher.flush();
-    }
-
-    // unsafe {
-    //     log::debug!(
-    //         "trigger kernel stack guard page :{:#x}",
-    //         (kstack_vaddr.data() - 8)
-    //     );
-    //     let guard_ptr = (kstack_vaddr.data() - 8) as *mut usize;
-    //     guard_ptr.write(0xfff); // Invalid
-    // }
-
-    // log::info!(
-    //     "[kernel stack alloc]: virt: {:#x}, phy: {:#x}",
-    //     kstack_vaddr.data(),
-    //     _kstack_paddr.data()
-    // );
-    (guard_vaddr, paddr)
-}
-
-unsafe fn dealloc_from_kernel_space(vaddr: VirtAddr, paddr: PhysAddr) {
-    use crate::arch::mm::kernel_page_flags;
-    use crate::arch::MMArch;
-    use crate::mm::allocator::page_frame::{deallocate_page_frames, PageFrameCount, PhysPageFrame};
-    use crate::mm::kernel_mapper::KernelMapper;
-    use crate::mm::MemoryManagementArch;
-
-    let _guard = KSTACK_LOCK.try_lock_irqsave().unwrap();
-
-    let need_size = KernelStack::SIZE * 2;
-    let page_num = PageFrameCount::new(need_size.div_ceil(MMArch::PAGE_SIZE).next_power_of_two());
-
-    // log::info!(
-    //     "[kernel stack dealloc]: virt: {:#x}, phy: {:#x}",
-    //     vaddr.data(),
-    //     paddr.data()
-    // );
-
-    let mut kernel_mapper = KernelMapper::lock();
-    let kernel_mapper = kernel_mapper.as_mut().unwrap();
-
-    // restore the guard page flags
-    for i in 0..KernelStack::SIZE / MMArch::PAGE_SIZE {
-        let guard_page_vaddr = vaddr + i * MMArch::PAGE_SIZE;
-        let flusher = kernel_mapper
-            .remap(guard_page_vaddr, kernel_page_flags(vaddr))
-            .unwrap();
-        flusher.flush();
-    }
-
-    // release the physical page
-    unsafe { deallocate_page_frames(PhysPageFrame::new(paddr), page_num) };
-}
-
-impl KernelStack {
-    pub const SIZE: usize = 0x8000;
-    pub const ALIGN: usize = 0x8000;
-
-    pub fn new() -> Result<Self, SystemError> {
-        if cfg!(feature = "kstack_protect") {
-            unsafe {
-                let (kstack_vaddr, kstack_paddr) = alloc_from_kernel_space();
-                let real_kstack_vaddr = kstack_vaddr + KernelStack::SIZE;
-                Ok(Self {
-                    stack: Some(
-                        AlignedBox::<[u8; KernelStack::SIZE], { KernelStack::ALIGN }>::new_unchecked(
-                            real_kstack_vaddr.data() as *mut [u8; KernelStack::SIZE],
-                        ),
-                    ),
-                    ty: KernelStackType::KernelSpace(kstack_vaddr, kstack_paddr),
-                })
-            }
-        } else {
-            Ok(Self {
-                stack: Some(
-                    AlignedBox::<[u8; KernelStack::SIZE], { KernelStack::ALIGN }>::new_zeroed()?,
-                ),
-                ty: KernelStackType::Dynamic,
-            })
-        }
-    }
-
-    /// 根据已有的空间，构造一个内核栈结构体
-    ///
-    /// 仅仅用于BSP启动时，为idle进程构造内核栈。其他时候使用这个函数，很可能造成错误！
-    pub unsafe fn from_existed(base: VirtAddr) -> Result<Self, SystemError> {
-        if base.is_null() || !base.check_aligned(Self::ALIGN) {
-            return Err(SystemError::EFAULT);
-        }
-
-        Ok(Self {
-            stack: Some(
-                AlignedBox::<[u8; KernelStack::SIZE], { KernelStack::ALIGN }>::new_unchecked(
-                    base.data() as *mut [u8; KernelStack::SIZE],
-                ),
-            ),
-            ty: KernelStackType::Static,
-        })
-    }
-
-    pub fn guard_page_address(&self) -> Option<VirtAddr> {
-        match self.ty {
-            KernelStackType::KernelSpace(kstack_virt_addr, _) => {
-                return Some(kstack_virt_addr);
-            }
-            _ => {
-                // 静态内核栈和动态内核栈没有guard page
-                return None;
-            }
-        }
-    }
-
-    pub fn guard_page_size(&self) -> Option<usize> {
-        match self.ty {
-            KernelStackType::KernelSpace(_, _) => {
-                return Some(KernelStack::SIZE);
-            }
-            _ => {
-                // 静态内核栈和动态内核栈没有guard page
-                return None;
-            }
-        }
-    }
-
-    /// 返回内核栈的起始虚拟地址(低地址)
-    pub fn start_address(&self) -> VirtAddr {
-        return VirtAddr::new(self.stack.as_ref().unwrap().as_ptr() as usize);
-    }
-
-    /// 返回内核栈的结束虚拟地址(高地址)(不包含该地址)
-    pub fn stack_max_address(&self) -> VirtAddr {
-        return VirtAddr::new(self.stack.as_ref().unwrap().as_ptr() as usize + Self::SIZE);
-    }
-
-    pub unsafe fn set_pcb(&mut self, pcb: Weak<ProcessControlBlock>) -> Result<(), SystemError> {
-        // 将一个Weak<ProcessControlBlock>放到内核栈的最低地址处
-        let p: *const ProcessControlBlock = Weak::into_raw(pcb);
-        let stack_bottom_ptr = self.start_address().data() as *mut *const ProcessControlBlock;
-
-        // 如果内核栈的最低地址处已经有了一个pcb，那么，这里就不再设置,直接返回错误
-        if unlikely(unsafe { !(*stack_bottom_ptr).is_null() }) {
-            error!("kernel stack bottom is not null: {:p}", *stack_bottom_ptr);
-            return Err(SystemError::EPERM);
-        }
-        // 将pcb的地址放到内核栈的最低地址处
-        unsafe {
-            *stack_bottom_ptr = p;
-        }
-
-        return Ok(());
-    }
-
-    /// 清除内核栈的pcb指针
-    ///
-    /// ## 参数
-    ///
-    /// - `force` : 如果为true,那么，即使该内核栈的pcb指针不为null，也会被强制清除而不处理Weak指针问题
-    pub unsafe fn clear_pcb(&mut self, force: bool) {
-        let stack_bottom_ptr = self.start_address().data() as *mut *const ProcessControlBlock;
-        if unlikely(unsafe { (*stack_bottom_ptr).is_null() }) {
-            return;
-        }
-
-        if !force {
-            let pcb_ptr: Weak<ProcessControlBlock> = Weak::from_raw(*stack_bottom_ptr);
-            drop(pcb_ptr);
-        }
-
-        *stack_bottom_ptr = core::ptr::null();
-    }
-
-    /// 返回指向当前内核栈pcb的Arc指针
-    #[allow(dead_code)]
-    pub unsafe fn pcb(&self) -> Option<Arc<ProcessControlBlock>> {
-        // 从内核栈的最低地址处取出pcb的地址
-        let p = self.stack.as_ref().unwrap().as_ptr() as *const *const ProcessControlBlock;
-        if unlikely(unsafe { (*p).is_null() }) {
-            return None;
-        }
-
-        // 为了防止内核栈的pcb指针被释放，这里需要将其包装一下，使得Arc的drop不会被调用
-        let weak_wrapper: ManuallyDrop<Weak<ProcessControlBlock>> =
-            ManuallyDrop::new(Weak::from_raw(*p));
-
-        let new_arc: Arc<ProcessControlBlock> = weak_wrapper.upgrade()?;
-        return Some(new_arc);
-    }
-}
-
-impl Drop for KernelStack {
-    fn drop(&mut self) {
-        if let Some(stack) = &self.stack {
-            let ptr = stack.as_ptr() as *const *const ProcessControlBlock;
-            if unsafe { !(*ptr).is_null() } {
-                let pcb_ptr: Weak<ProcessControlBlock> = unsafe { Weak::from_raw(*ptr) };
-                drop(pcb_ptr);
-            }
-        }
-        match self.ty {
-            KernelStackType::KernelSpace(kstack_virt_addr, kstack_phy_addr) => {
-                // 释放内核栈
-                unsafe {
-                    dealloc_from_kernel_space(kstack_virt_addr, kstack_phy_addr);
-                }
-                let bx = self.stack.take();
-                core::mem::forget(bx);
-            }
-            KernelStackType::Static => {
-                let bx = self.stack.take();
-                core::mem::forget(bx);
-            }
-            KernelStackType::Dynamic => {}
-        }
-    }
-}
-
-pub fn process_init() {
-    ProcessManager::init();
-}
-
-#[derive(Debug)]
-pub struct ProcessSignalInfo {
-    // 当前进程被屏蔽的信号
-    sig_blocked: SigSet,
-    // 暂存旧信号，用于恢复
-    saved_sigmask: SigSet,
-    // sig_pending 中存储当前线程要处理的信号
-    sig_pending: SigPending,
-
-    // 当前进程对应的tty
-    tty: Option<Arc<TtyCore>>,
-    has_child_subreaper: bool,
-
-    /// 标记当前进程是否是一个“子进程收割者”
-    ///
-    /// todo: 在prctl里面实现设置这个标志位的功能
-    is_child_subreaper: bool,
-
-    /// boolean value for session group leader
-    pub is_session_leader: bool,
-
-    /// OOM killer score adjustment exposed through `/proc/[pid]/oom_score_adj`.
-    oom_score_adj: i16,
-    /// Minimum oom_score_adj an unprivileged writer may set.
-    ///
-    /// Linux updates this value only from CAP_SYS_RESOURCE writes. The default
-    /// is 0, so an unprivileged process cannot make itself more protected than
-    /// the initial state.
-    oom_score_adj_min: i16,
-}
-
-impl ProcessSignalInfo {
-    pub fn sig_blocked(&self) -> &SigSet {
-        &self.sig_blocked
-    }
-
-    pub fn sig_pending(&self) -> &SigPending {
-        &self.sig_pending
-    }
-
-    pub fn sig_pending_mut(&mut self) -> &mut SigPending {
-        &mut self.sig_pending
-    }
-
-    pub fn sig_block_mut(&mut self) -> &mut SigSet {
-        &mut self.sig_blocked
-    }
-
-    pub fn saved_sigmask(&self) -> &SigSet {
-        &self.saved_sigmask
-    }
-
-    pub fn saved_sigmask_mut(&mut self) -> &mut SigSet {
-        &mut self.saved_sigmask
-    }
-
-    pub fn tty(&self) -> Option<Arc<TtyCore>> {
-        self.tty.clone()
-    }
-
-    pub fn set_tty(&mut self, tty: Option<Arc<TtyCore>>) {
-        self.tty = tty;
-    }
-
-    /// 从当前线程的 pending 中取出下一个要处理的信号
-    ///
-    /// ## 参数
-    ///
-    /// - `sig_mask` 被忽略掉的信号
-    pub fn dequeue_thread_signal(&mut self, sig_mask: &SigSet) -> (Signal, Option<SigInfo>) {
-        self.sig_pending.dequeue_signal(sig_mask)
-    }
-
-    pub fn has_child_subreaper(&self) -> bool {
-        self.has_child_subreaper
-    }
-
-    pub fn set_has_child_subreaper(&mut self, has_child_subreaper: bool) {
-        self.has_child_subreaper = has_child_subreaper;
-    }
-
-    pub fn is_child_subreaper(&self) -> bool {
-        self.is_child_subreaper
-    }
-
-    pub fn set_is_child_subreaper(&mut self, is_child_subreaper: bool) {
-        self.is_child_subreaper = is_child_subreaper;
-    }
-
-    pub fn oom_score_adj(&self) -> i16 {
-        self.oom_score_adj
-    }
-
-    pub fn set_oom_score_adj(&mut self, oom_score_adj: i16) {
-        self.oom_score_adj = oom_score_adj;
-    }
-
-    pub fn oom_score_adj_min(&self) -> i16 {
-        self.oom_score_adj_min
-    }
-
-    pub fn set_oom_score_adj_min(&mut self, oom_score_adj_min: i16) {
-        self.oom_score_adj_min = oom_score_adj_min;
-    }
-}
-
-impl Default for ProcessSignalInfo {
-    fn default() -> Self {
-        Self {
-            sig_blocked: SigSet::empty(),
-            saved_sigmask: SigSet::empty(),
-            sig_pending: SigPending::default(),
-            tty: None,
-            has_child_subreaper: false,
-            is_child_subreaper: false,
-            is_session_leader: false,
-            oom_score_adj: 0,
-            oom_score_adj_min: 0,
-        }
-    }
 }
